@@ -1,8 +1,25 @@
-module github.com/gofiber/fiber/v2
+module github.com/gofiber/fiber/v3
 
-go 1.14
+go 1.20
 
 require (
-	github.com/valyala/fasthttp v1.17.0
-	golang.org/x/sys v0.0.0-20201119102817-f84b799fce68
+	github.com/gofiber/utils/v2 v2.0.0-beta.3
+	github.com/google/uuid v1.6.0
+	github.com/mattn/go-colorable v0.1.13
+	github.com/mattn/go-isatty v0.0.20
+	github.com/stretchr/testify v1.8.4
+	github.com/tinylib/msgp v1.1.8
+	github.com/valyala/bytebufferpool v1.0.0
+	github.com/valyala/fasthttp v1.51.0
+)
+
+require (
+	github.com/andybalholm/brotli v1.1.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/klauspost/compress v1.17.4 // indirect
+	github.com/philhofer/fwd v1.1.2 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/valyala/tcplisten v1.0.0 // indirect
+	golang.org/x/sys v0.16.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
