@@ -2,54 +2,53 @@
 // 🤖 Github Repository: https://github.com/gofiber/fiber
 // 📌 API Documentation: https://docs.gofiber.io
 
-// Package fiber
-// Fiber is an Express inspired web framework built on top of Fasthttp,
+// Package fiber is an Express inspired web framework built on top of Fasthttp,
 // the fastest HTTP engine for Go. Designed to ease things up for fast
 // development with zero memory allocation and performance in mind.
-
 package fiber
 
 import (
 	"bufio"
+	"context"
+	"encoding/json"
+	"encoding/xml"
+	"errors"
 	"fmt"
 	"net"
 	"net/http"
 	"net/http/httputil"
-	"os"
 	"reflect"
-	"runtime"
 	"strconv"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/gofiber/fiber/v2/internal/colorable"
-	"github.com/gofiber/fiber/v2/internal/isatty"
-	"github.com/gofiber/fiber/v2/utils"
+	"github.com/gofiber/fiber/v3/log"
+	"github.com/gofiber/utils/v2"
 
 	"github.com/valyala/fasthttp"
 )
 
 // Version of current fiber package
-const Version = "2.2.1"
+const Version = "3.0.0-beta.1"
 
 // Handler defines a function to serve HTTP requests.
-type Handler = func(*Ctx) error
+type Handler = func(Ctx) error
 
-// Map is a shortcut for map[string]interface{}, useful for JSON returns
-type Map map[string]interface{}
+// Map is a shortcut for map[string]any, useful for JSON returns
+type Map map[string]any
 
 // Storage interface for communicating with different database/key-value
 // providers
 type Storage interface {
 	// Get gets the value for the given key.
-	// It returns ErrNotFound if the storage does not contain the key.
+	// `nil, nil` is returned when the key does not exist
 	Get(key string) ([]byte, error)
 
-	// Set stores the given value for the given key along with a
-	// time-to-live expiration value, 0 means live for ever
+	// Set stores the given value for the given key along
+	// with an expiration value, 0 means no expiration.
 	// Empty key or value will be ignored without an error.
-	Set(key string, val []byte, ttl time.Duration) error
+	Set(key string, val []byte, exp time.Duration) error
 
 	// Delete deletes the value for the given key.
 	// It returns no error if the storage does not contain the key,
@@ -65,17 +64,19 @@ type Storage interface {
 
 // ErrorHandler defines a function that will process all errors
 // returned from any handlers in the stack
-//  cfg := fiber.Config{}
-//  cfg.ErrorHandler = func(c *Ctx, err error) error {
-//   code := StatusInternalServerError
-//   if e, ok := err.(*Error); ok {
-//     code = e.Code
-//   }
-//   c.Set(HeaderContentType, MIMETextPlainCharsetUTF8)
-//   return c.Status(code).SendString(err.Error())
-//  }
-//  app := fiber.New(cfg)
-type ErrorHandler = func(*Ctx, error) error
+//
+//	cfg := fiber.Config{}
+//	cfg.ErrorHandler = func(c Ctx, err error) error {
+//	 code := StatusInternalServerError
+//	 var e *fiber.Error
+//	 if errors.As(err, &e) {
+//	   code = e.Code
+//	 }
+//	 c.Set(HeaderContentType, MIMETextPlainCharsetUTF8)
+//	 return c.Status(code).SendString(err.Error())
+//	}
+//	app := fiber.New(cfg)
+type ErrorHandler = func(Ctx, error) error
 
 // Error represents an error that occurred while handling a request.
 type Error struct {
@@ -90,25 +91,42 @@ type App struct {
 	stack [][]*Route
 	// Route stack divided by HTTP methods and route prefixes
 	treeStack []map[string][]*Route
+	// contains the information if the route stack has been changed to build the optimized tree
+	routesRefreshed bool
 	// Amount of registered routes
-	routesCount int
+	routesCount uint32
 	// Amount of registered handlers
-	handlerCount int
+	handlersCount uint32
 	// Ctx pool
 	pool sync.Pool
 	// Fasthttp server
 	server *fasthttp.Server
 	// App config
 	config Config
+	// Converts string to a byte slice
+	getBytes func(s string) (b []byte)
+	// Converts byte slice to a string
+	getString func(b []byte) string
+	// Hooks
+	hooks *Hooks
+	// Latest route & group
+	latestRoute *Route
+	// newCtxFunc
+	newCtxFunc func(app *App) CustomCtx
+	// custom binders
+	customBinders []CustomBinder
+	// TLS handler
+	tlsHandler *TLSHandler
+	// Mount fields
+	mountFields *mountFields
+	// Indicates if the value was explicitly configured
+	configured Config
+	// customConstraints is a list of external constraints
+	customConstraints []CustomConstraint
 }
 
 // Config is a struct holding the server settings.
 type Config struct {
-	// When set to true, this will spawn multiple Go processes listening on the same port.
-	//
-	// Default: false
-	Prefork bool `json:"prefork"`
-
 	// Enables the "Server: value" HTTP header.
 	//
 	// Default: ""
@@ -136,17 +154,12 @@ type Config struct {
 	Immutable bool `json:"immutable"`
 
 	// When set to true, converts all encoded characters in the route back
-	// before setting the path for the context, so that the routing can also
-	// work with urlencoded special characters.
+	// before setting the path for the context, so that the routing,
+	// the returning of the current url from the context `ctx.Path()`
+	// and the parameters `ctx.Params(%key%)` with decoded characters will work
 	//
 	// Default: false
 	UnescapePath bool `json:"unescape_path"`
-
-	// Enable or disable ETag header generation, since both weak and strong etags are generated
-	// using the same hashing method (CRC-32). Weak ETags are the default when enabled.
-	//
-	// Default: false
-	ETag bool `json:"etag"`
 
 	// Max body size that the server accepts.
 	// -1 will decline any body size
@@ -163,6 +176,16 @@ type Config struct {
 	//
 	// Default: nil
 	Views Views `json:"-"`
+
+	// Views Layout is the global layout for all template render until override on Render function.
+	//
+	// Default: ""
+	ViewsLayout string `json:"views_layout"`
+
+	// PassLocalsToViews Enables passing of the locals set on a fiber.Ctx to the template engine
+	//
+	// Default: false
+	PassLocalsToViews bool `json:"pass_locals_to_views"`
 
 	// The amount of time allowed to read the full request including body.
 	// It is reset after the request handler has returned.
@@ -245,10 +268,23 @@ type Config struct {
 	// Default: false
 	DisableHeaderNormalizing bool `json:"disable_header_normalizing"`
 
-	// When set to true, it will not print out the «Fiber» ASCII art and listening address.
+	// This function allows to setup app name for the app
 	//
-	// Default: false
-	DisableStartupMessage bool `json:"disable_startup_message"`
+	// Default: nil
+	AppName string `json:"app_name"`
+
+	// StreamRequestBody enables request body streaming,
+	// and calls the handler sooner when given body is
+	// larger then the current limit.
+	StreamRequestBody bool
+
+	// Will not pre parse Multipart Form data if set to true.
+	//
+	// This option is useful for servers that desire to treat
+	// multipart form data as a binary blob, or choose when to parse the data.
+	//
+	// Server pre parses multipart form data by default.
+	DisablePreParseMultipartForm bool
 
 	// Aggressively reduces memory usage at the cost of higher CPU usage
 	// if set to true.
@@ -262,11 +298,91 @@ type Config struct {
 
 	// FEATURE: v2.3.x
 	// The router executes the same handler by default if StrictRouting or CaseSensitive is disabled.
-	// Enabling RedirectFixedPath will change this behaviour into a client redirect to the original route path.
+	// Enabling RedirectFixedPath will change this behavior into a client redirect to the original route path.
 	// Using the status code 301 for GET requests and 308 for all other request methods.
 	//
 	// Default: false
 	// RedirectFixedPath bool
+
+	// When set by an external client of Fiber it will use the provided implementation of a
+	// JSONMarshal
+	//
+	// Allowing for flexibility in using another json library for encoding
+	// Default: json.Marshal
+	JSONEncoder utils.JSONMarshal `json:"-"`
+
+	// When set by an external client of Fiber it will use the provided implementation of a
+	// JSONUnmarshal
+	//
+	// Allowing for flexibility in using another json library for decoding
+	// Default: json.Unmarshal
+	JSONDecoder utils.JSONUnmarshal `json:"-"`
+
+	// XMLEncoder set by an external client of Fiber it will use the provided implementation of a
+	// XMLMarshal
+	//
+	// Allowing for flexibility in using another XML library for encoding
+	// Default: xml.Marshal
+	XMLEncoder utils.XMLMarshal `json:"-"`
+
+	// If you find yourself behind some sort of proxy, like a load balancer,
+	// then certain header information may be sent to you using special X-Forwarded-* headers or the Forwarded header.
+	// For example, the Host HTTP header is usually used to return the requested host.
+	// But when you’re behind a proxy, the actual host may be stored in an X-Forwarded-Host header.
+	//
+	// If you are behind a proxy, you should enable TrustedProxyCheck to prevent header spoofing.
+	// If you enable EnableTrustedProxyCheck and leave TrustedProxies empty Fiber will skip
+	// all headers that could be spoofed.
+	// If request ip in TrustedProxies whitelist then:
+	//   1. c.Scheme() get value from X-Forwarded-Proto, X-Forwarded-Protocol, X-Forwarded-Ssl or X-Url-Scheme header
+	//   2. c.IP() get value from ProxyHeader header.
+	//   3. c.Host() and c.Hostname() get value from X-Forwarded-Host header
+	// But if request ip NOT in Trusted Proxies whitelist then:
+	//   1. c.Scheme() WON't get value from X-Forwarded-Proto, X-Forwarded-Protocol, X-Forwarded-Ssl or X-Url-Scheme header,
+	//    will return https in case when tls connection is handled by the app, of http otherwise
+	//   2. c.IP() WON'T get value from ProxyHeader header, will return RemoteIP() from fasthttp context
+	//   3. c.Host() and c.Hostname() WON'T get value from X-Forwarded-Host header, fasthttp.Request.URI().Host()
+	//    will be used to get the hostname.
+	//
+	// Default: false
+	EnableTrustedProxyCheck bool `json:"enable_trusted_proxy_check"`
+
+	// Read EnableTrustedProxyCheck doc.
+	//
+	// Default: []string
+	TrustedProxies     []string `json:"trusted_proxies"`
+	trustedProxiesMap  map[string]struct{}
+	trustedProxyRanges []*net.IPNet
+
+	// If set to true, c.IP() and c.IPs() will validate IP addresses before returning them.
+	// Also, c.IP() will return only the first valid IP rather than just the raw header
+	// WARNING: this has a performance cost associated with it.
+	//
+	// Default: false
+	EnableIPValidation bool `json:"enable_ip_validation"`
+
+	// You can define custom color scheme. They'll be used for startup message, route list and some middlewares.
+	//
+	// Optional. Default: DefaultColors
+	ColorScheme Colors `json:"color_scheme"`
+
+	// If you want to validate header/form/query... automatically when to bind, you can define struct validator.
+	// Fiber doesn't have default validator, so it'll skip validator step if you don't use any validator.
+	//
+	// Default: nil
+	StructValidator StructValidator
+
+	// RequestMethods provides customizibility for HTTP methods. You can add/remove methods as you wish.
+	//
+	// Optional. Default: DefaultMethods
+	RequestMethods []string
+
+	// EnableSplittingOnParsers splits the query/body/header parameters by comma when it's true.
+	// For example, you can use it to parse multiple values from a query parameter like this:
+	//   /api?foo=bar,baz == foo[]=bar&foo[]=baz
+	//
+	// Optional. Default: false
+	EnableSplittingOnParsers bool `json:"enable_splitting_on_parsers"`
 }
 
 // Static defines configuration options when defining static assets.
@@ -284,6 +400,10 @@ type Static struct {
 	// Optional. Default value false.
 	Browse bool `json:"browse"`
 
+	// When set to true, enables direct download.
+	// Optional. Default value false.
+	Download bool `json:"download"`
+
 	// The name of the index file for serving a directory.
 	// Optional. Default value "index.html".
 	Index string `json:"index"`
@@ -299,6 +419,24 @@ type Static struct {
 	//
 	// Optional. Default value 0.
 	MaxAge int `json:"max_age"`
+
+	// ModifyResponse defines a function that allows you to alter the response.
+	//
+	// Optional. Default: nil
+	ModifyResponse Handler
+
+	// Next defines a function to skip this middleware when returned true.
+	//
+	// Optional. Default: nil
+	Next func(c Ctx) bool
+}
+
+// RouteMessage is some message need to be print when server starts
+type RouteMessage struct {
+	name     string
+	method   string
+	path     string
+	handlers string
 }
 
 // Default Config values
@@ -310,10 +448,24 @@ const (
 	DefaultCompressedFileSuffix = ".fiber.gz"
 )
 
+// HTTP methods enabled by default
+var DefaultMethods = []string{
+	MethodGet,
+	MethodHead,
+	MethodPost,
+	MethodPut,
+	MethodDelete,
+	MethodConnect,
+	MethodOptions,
+	MethodTrace,
+	MethodPatch,
+}
+
 // DefaultErrorHandler that process return errors from handlers
-var DefaultErrorHandler = func(c *Ctx, err error) error {
+func DefaultErrorHandler(c Ctx, err error) error {
 	code := StatusInternalServerError
-	if e, ok := err.(*Error); ok {
+	var e *Error
+	if errors.As(err, &e) {
 		code = e.Code
 	}
 	c.Set(HeaderContentType, MIMETextPlainCharsetUTF8)
@@ -321,37 +473,46 @@ var DefaultErrorHandler = func(c *Ctx, err error) error {
 }
 
 // New creates a new Fiber named instance.
-//  app := fiber.New()
+//
+//	app := fiber.New()
+//
 // You can pass optional configuration options by passing a Config struct:
-//  app := fiber.New(fiber.Config{
-//      Prefork: true,
-//      ServerHeader: "Fiber",
-//  })
+//
+//	app := fiber.New(fiber.Config{
+//	    Prefork: true,
+//	    ServerHeader: "Fiber",
+//	})
 func New(config ...Config) *App {
 	// Create a new app
 	app := &App{
-		// Create router stack
-		stack:     make([][]*Route, len(intMethod)),
-		treeStack: make([]map[string][]*Route, len(intMethod)),
-		// Create Ctx pool
-		pool: sync.Pool{
-			New: func() interface{} {
-				return new(Ctx)
-			},
-		},
 		// Create config
-		config: Config{},
+		config:        Config{},
+		getBytes:      utils.UnsafeBytes,
+		getString:     utils.UnsafeString,
+		latestRoute:   &Route{},
+		customBinders: []CustomBinder{},
 	}
+
+	// Create Ctx pool
+	app.pool = sync.Pool{
+		New: func() any {
+			return app.NewCtx(&fasthttp.RequestCtx{})
+		},
+	}
+
+	// Define hooks
+	app.hooks = newHooks(app)
+
+	// Define mountFields
+	app.mountFields = newMountFields(app)
+
 	// Override config if provided
 	if len(config) > 0 {
 		app.config = config[0]
 	}
 
-	if app.config.ETag {
-		if !IsChild() {
-			fmt.Println("[Warning] Config.ETag is deprecated since v2.0.6, please use 'middleware/etag'.")
-		}
-	}
+	// Initialize configured before defaults are set
+	app.configured = app.config
 
 	// Override default values
 	if app.config.BodyLimit == 0 {
@@ -370,11 +531,37 @@ func New(config ...Config) *App {
 		app.config.CompressedFileSuffix = DefaultCompressedFileSuffix
 	}
 	if app.config.Immutable {
-		getBytes, getString = getBytesImmutable, getStringImmutable
+		app.getBytes, app.getString = getBytesImmutable, getStringImmutable
 	}
+
 	if app.config.ErrorHandler == nil {
 		app.config.ErrorHandler = DefaultErrorHandler
 	}
+
+	if app.config.JSONEncoder == nil {
+		app.config.JSONEncoder = json.Marshal
+	}
+	if app.config.JSONDecoder == nil {
+		app.config.JSONDecoder = json.Unmarshal
+	}
+	if app.config.XMLEncoder == nil {
+		app.config.XMLEncoder = xml.Marshal
+	}
+	if len(app.config.RequestMethods) == 0 {
+		app.config.RequestMethods = DefaultMethods
+	}
+
+	app.config.trustedProxiesMap = make(map[string]struct{}, len(app.config.TrustedProxies))
+	for _, ipAddress := range app.config.TrustedProxies {
+		app.handleTrustedProxy(ipAddress)
+	}
+
+	// Create router stack
+	app.stack = make([][]*Route, len(app.config.RequestMethods))
+	app.treeStack = make([]map[string][]*Route, len(app.config.RequestMethods))
+
+	// Override colors
+	app.config.ColorScheme = defaultColors(app.config.ColorScheme)
 
 	// Init app
 	app.init()
@@ -383,131 +570,255 @@ func New(config ...Config) *App {
 	return app
 }
 
-// Mount attaches another app instance as a subrouter along a routing path.
-// It's very useful to split up a large API as many independent routers and
-// compose them as a single service using Mount.
-func (app *App) Mount(prefix string, fiber *App) Router {
-	stack := fiber.Stack()
-	for m := range stack {
-		for r := range stack[m] {
-			route := app.copyRoute(stack[m][r])
-			app.addRoute(route.Method, app.addPrefixToRoute(prefix, route))
+// Adds an ip address to trustedProxyRanges or trustedProxiesMap based on whether it is an IP range or not
+func (app *App) handleTrustedProxy(ipAddress string) {
+	if strings.Contains(ipAddress, "/") {
+		_, ipNet, err := net.ParseCIDR(ipAddress)
+		if err != nil {
+			log.Warnf("IP range %q could not be parsed: %v", ipAddress, err)
+		} else {
+			app.config.trustedProxyRanges = append(app.config.trustedProxyRanges, ipNet)
+		}
+	} else {
+		app.config.trustedProxiesMap[ipAddress] = struct{}{}
+	}
+}
+
+// NewCtxFunc allows to customize ctx methods as we want.
+// Note: It doesn't allow adding new methods, only customizing exist methods.
+func (app *App) NewCtxFunc(function func(app *App) CustomCtx) {
+	app.newCtxFunc = function
+}
+
+// RegisterCustomConstraint allows to register custom constraint.
+func (app *App) RegisterCustomConstraint(constraint CustomConstraint) {
+	app.customConstraints = append(app.customConstraints, constraint)
+}
+
+// You can register custom binders to use as Bind().Custom("name").
+// They should be compatible with CustomBinder interface.
+func (app *App) RegisterCustomBinder(binder CustomBinder) {
+	app.customBinders = append(app.customBinders, binder)
+}
+
+// You can use SetTLSHandler to use ClientHelloInfo when using TLS with Listener.
+func (app *App) SetTLSHandler(tlsHandler *TLSHandler) {
+	// Attach the tlsHandler to the config
+	app.mutex.Lock()
+	app.tlsHandler = tlsHandler
+	app.mutex.Unlock()
+}
+
+// Name Assign name to specific route.
+func (app *App) Name(name string) Router {
+	app.mutex.Lock()
+	defer app.mutex.Unlock()
+
+	for _, routes := range app.stack {
+		for _, route := range routes {
+			isMethodValid := route.Method == app.latestRoute.Method || app.latestRoute.use ||
+				(app.latestRoute.Method == MethodGet && route.Method == MethodHead)
+
+			if route.Path == app.latestRoute.Path && isMethodValid {
+				route.Name = name
+				if route.group != nil {
+					route.Name = route.group.name + route.Name
+				}
+			}
 		}
 	}
+
+	if err := app.hooks.executeOnNameHooks(*app.latestRoute); err != nil {
+		panic(err)
+	}
+
 	return app
+}
+
+// GetRoute Get route by name
+func (app *App) GetRoute(name string) Route {
+	for _, routes := range app.stack {
+		for _, route := range routes {
+			if route.Name == name {
+				return *route
+			}
+		}
+	}
+
+	return Route{}
+}
+
+// GetRoutes Get all routes. When filterUseOption equal to true, it will filter the routes registered by the middleware.
+func (app *App) GetRoutes(filterUseOption ...bool) []Route {
+	var rs []Route
+	var filterUse bool
+	if len(filterUseOption) != 0 {
+		filterUse = filterUseOption[0]
+	}
+	for _, routes := range app.stack {
+		for _, route := range routes {
+			if filterUse && route.use {
+				continue
+			}
+			rs = append(rs, *route)
+		}
+	}
+	return rs
 }
 
 // Use registers a middleware route that will match requests
 // with the provided prefix (which is optional and defaults to "/").
+// Also, you can pass another app instance as a sub-router along a routing path.
+// It's very useful to split up a large API as many independent routers and
+// compose them as a single service using Use. The fiber's error handler and
+// any of the fiber's sub apps are added to the application's error handlers
+// to be invoked on errors that happen within the prefix route.
 //
-//  app.Use(func(c *fiber.Ctx) error {
-//       return c.Next()
-//  })
-//  app.Use("/api", func(c *fiber.Ctx) error {
-//       return c.Next()
-//  })
-//  app.Use("/api", handler, func(c *fiber.Ctx) error {
-//       return c.Next()
-//  })
+//		app.Use(func(c fiber.Ctx) error {
+//		     return c.Next()
+//		})
+//		app.Use("/api", func(c fiber.Ctx) error {
+//		     return c.Next()
+//		})
+//		app.Use("/api", handler, func(c fiber.Ctx) error {
+//		     return c.Next()
+//		})
+//	 	subApp := fiber.New()
+//		app.Use("/mounted-path", subApp)
 //
 // This method will match all HTTP verbs: GET, POST, PUT, HEAD etc...
-func (app *App) Use(args ...interface{}) Router {
+func (app *App) Use(args ...any) Router {
 	var prefix string
+	var subApp *App
+	var prefixes []string
 	var handlers []Handler
 
 	for i := 0; i < len(args); i++ {
 		switch arg := args[i].(type) {
 		case string:
 			prefix = arg
+		case *App:
+			subApp = arg
+		case []string:
+			prefixes = arg
 		case Handler:
 			handlers = append(handlers, arg)
 		default:
 			panic(fmt.Sprintf("use: invalid handler %v\n", reflect.TypeOf(arg)))
 		}
 	}
-	app.register(methodUse, prefix, handlers...)
+
+	if len(prefixes) == 0 {
+		prefixes = append(prefixes, prefix)
+	}
+
+	for _, prefix := range prefixes {
+		if subApp != nil {
+			app.mount(prefix, subApp)
+			return app
+		}
+
+		app.register([]string{methodUse}, prefix, nil, nil, handlers...)
+	}
+
 	return app
 }
 
 // Get registers a route for GET methods that requests a representation
 // of the specified resource. Requests using GET should only retrieve data.
-func (app *App) Get(path string, handlers ...Handler) Router {
-	return app.Add(MethodHead, path, handlers...).Add(MethodGet, path, handlers...)
+func (app *App) Get(path string, handler Handler, middleware ...Handler) Router {
+	return app.Add([]string{MethodGet}, path, handler, middleware...)
 }
 
 // Head registers a route for HEAD methods that asks for a response identical
 // to that of a GET request, but without the response body.
-func (app *App) Head(path string, handlers ...Handler) Router {
-	return app.Add(MethodHead, path, handlers...)
+func (app *App) Head(path string, handler Handler, middleware ...Handler) Router {
+	return app.Add([]string{MethodHead}, path, handler, middleware...)
 }
 
 // Post registers a route for POST methods that is used to submit an entity to the
 // specified resource, often causing a change in state or side effects on the server.
-func (app *App) Post(path string, handlers ...Handler) Router {
-	return app.Add(MethodPost, path, handlers...)
+func (app *App) Post(path string, handler Handler, middleware ...Handler) Router {
+	return app.Add([]string{MethodPost}, path, handler, middleware...)
 }
 
 // Put registers a route for PUT methods that replaces all current representations
 // of the target resource with the request payload.
-func (app *App) Put(path string, handlers ...Handler) Router {
-	return app.Add(MethodPut, path, handlers...)
+func (app *App) Put(path string, handler Handler, middleware ...Handler) Router {
+	return app.Add([]string{MethodPut}, path, handler, middleware...)
 }
 
 // Delete registers a route for DELETE methods that deletes the specified resource.
-func (app *App) Delete(path string, handlers ...Handler) Router {
-	return app.Add(MethodDelete, path, handlers...)
+func (app *App) Delete(path string, handler Handler, middleware ...Handler) Router {
+	return app.Add([]string{MethodDelete}, path, handler, middleware...)
 }
 
 // Connect registers a route for CONNECT methods that establishes a tunnel to the
 // server identified by the target resource.
-func (app *App) Connect(path string, handlers ...Handler) Router {
-	return app.Add(MethodConnect, path, handlers...)
+func (app *App) Connect(path string, handler Handler, middleware ...Handler) Router {
+	return app.Add([]string{MethodConnect}, path, handler, middleware...)
 }
 
 // Options registers a route for OPTIONS methods that is used to describe the
 // communication options for the target resource.
-func (app *App) Options(path string, handlers ...Handler) Router {
-	return app.Add(MethodOptions, path, handlers...)
+func (app *App) Options(path string, handler Handler, middleware ...Handler) Router {
+	return app.Add([]string{MethodOptions}, path, handler, middleware...)
 }
 
 // Trace registers a route for TRACE methods that performs a message loop-back
 // test along the path to the target resource.
-func (app *App) Trace(path string, handlers ...Handler) Router {
-	return app.Add(MethodTrace, path, handlers...)
+func (app *App) Trace(path string, handler Handler, middleware ...Handler) Router {
+	return app.Add([]string{MethodTrace}, path, handler, middleware...)
 }
 
 // Patch registers a route for PATCH methods that is used to apply partial
 // modifications to a resource.
-func (app *App) Patch(path string, handlers ...Handler) Router {
-	return app.Add(MethodPatch, path, handlers...)
+func (app *App) Patch(path string, handler Handler, middleware ...Handler) Router {
+	return app.Add([]string{MethodPatch}, path, handler, middleware...)
 }
 
-// Add allows you to specify a HTTP method to register a route
-func (app *App) Add(method, path string, handlers ...Handler) Router {
-	return app.register(method, path, handlers...)
+// Add allows you to specify multiple HTTP methods to register a route.
+func (app *App) Add(methods []string, path string, handler Handler, middleware ...Handler) Router {
+	app.register(methods, path, nil, handler, middleware...)
+
+	return app
 }
 
 // Static will create a file server serving static files
 func (app *App) Static(prefix, root string, config ...Static) Router {
-	return app.registerStatic(prefix, root, config...)
-}
+	app.registerStatic(prefix, root, config...)
 
-// All will register the handler on all HTTP methods
-func (app *App) All(path string, handlers ...Handler) Router {
-	for _, method := range intMethod {
-		_ = app.Add(method, path, handlers...)
-	}
 	return app
 }
 
+// All will register the handler on all HTTP methods
+func (app *App) All(path string, handler Handler, middleware ...Handler) Router {
+	return app.Add(app.config.RequestMethods, path, handler, middleware...)
+}
+
 // Group is used for Routes with common prefix to define a new sub-router with optional middleware.
-//  api := app.Group("/api")
-//  api.Get("/users", handler)
+//
+//	api := app.Group("/api")
+//	api.Get("/users", handler)
 func (app *App) Group(prefix string, handlers ...Handler) Router {
+	grp := &Group{Prefix: prefix, app: app}
 	if len(handlers) > 0 {
-		app.register(methodUse, prefix, handlers...)
+		app.register([]string{methodUse}, prefix, grp, nil, handlers...)
 	}
-	return &Group{prefix: prefix, app: app}
+	if err := app.hooks.executeOnGroupHooks(*grp); err != nil {
+		panic(err)
+	}
+
+	return grp
+}
+
+// Route is used to define routes with a common prefix inside the common function.
+// Uses Group method to define new sub-router.
+func (app *App) Route(path string) Register {
+	// Create new route
+	route := &Registering{app: app, path: path}
+
+	return route
 }
 
 // Error makes it compatible with the `error` interface.
@@ -517,54 +828,14 @@ func (e *Error) Error() string {
 
 // NewError creates a new Error instance with an optional message
 func NewError(code int, message ...string) *Error {
-	e := &Error{
-		Code: code,
+	err := &Error{
+		Code:    code,
+		Message: utils.StatusMessage(code),
 	}
 	if len(message) > 0 {
-		e.Message = message[0]
-	} else {
-		e.Message = utils.StatusMessage(code)
+		err.Message = message[0]
 	}
-	return e
-}
-
-// Listener can be used to pass a custom listener.
-func (app *App) Listener(ln net.Listener) error {
-	// Prefork is supported for custom listeners
-	if app.config.Prefork {
-		addr, tls := lnMetadata(ln)
-		return app.prefork(addr, tls)
-	}
-
-	// Print startup message
-	if !app.config.DisableStartupMessage {
-		app.startupMessage(ln.Addr().String(), false, "")
-	}
-
-	// TODO: Detect TLS
-	return app.server.Serve(ln)
-}
-
-// Listen serves HTTP requests from the given addr.
-//
-//  app.Listen(":8080")
-//  app.Listen("127.0.0.1:8080")
-func (app *App) Listen(addr string) error {
-	// Start prefork
-	if app.config.Prefork {
-		return app.prefork(addr, nil)
-	}
-	// Setup listener
-	ln, err := net.Listen("tcp4", addr)
-	if err != nil {
-		return err
-	}
-	// Print startup message
-	if !app.config.DisableStartupMessage {
-		app.startupMessage(ln.Addr().String(), false, "")
-	}
-	// Start listening
-	return app.server.Serve(ln)
+	return err
 }
 
 // Config returns the app config as value ( read-only ).
@@ -573,8 +844,10 @@ func (app *App) Config() Config {
 }
 
 // Handler returns the server handler.
-func (app *App) Handler() fasthttp.RequestHandler {
-	return app.handler
+func (app *App) Handler() fasthttp.RequestHandler { //revive:disable-line:confusing-naming // Having both a Handler() (uppercase) and a handler() (lowercase) is fine. TODO: Use nolint:revive directive instead. See https://github.com/golangci/golangci-lint/issues/3476
+	// prepare the server for the start
+	app.startupProcess()
+	return app.requestHandler
 }
 
 // Stack returns the raw router stack.
@@ -582,19 +855,51 @@ func (app *App) Stack() [][]*Route {
 	return app.stack
 }
 
+// HandlersCount returns the amount of registered handlers.
+func (app *App) HandlersCount() uint32 {
+	return app.handlersCount
+}
+
 // Shutdown gracefully shuts down the server without interrupting any active connections.
-// Shutdown works by first closing all open listeners and then waiting indefinitely for all connections to return to idle and then shut down.
+// Shutdown works by first closing all open listeners and then waiting indefinitely for all connections to return to idle before shutting down.
 //
 // Make sure the program doesn't exit and waits instead for Shutdown to return.
 //
 // Shutdown does not close keepalive connections so its recommended to set ReadTimeout to something else than 0.
 func (app *App) Shutdown() error {
+	return app.ShutdownWithContext(context.Background())
+}
+
+// ShutdownWithTimeout gracefully shuts down the server without interrupting any active connections. However, if the timeout is exceeded,
+// ShutdownWithTimeout will forcefully close any active connections.
+// ShutdownWithTimeout works by first closing all open listeners and then waiting for all connections to return to idle before shutting down.
+//
+// Make sure the program doesn't exit and waits instead for ShutdownWithTimeout to return.
+//
+// ShutdownWithTimeout does not close keepalive connections so its recommended to set ReadTimeout to something else than 0.
+func (app *App) ShutdownWithTimeout(timeout time.Duration) error {
+	ctx, cancelFunc := context.WithTimeout(context.Background(), timeout)
+	defer cancelFunc()
+	return app.ShutdownWithContext(ctx)
+}
+
+// ShutdownWithContext shuts down the server including by force if the context's deadline is exceeded.
+//
+// Make sure the program doesn't exit and waits instead for ShutdownWithTimeout to return.
+//
+// ShutdownWithContext does not close keepalive connections so its recommended to set ReadTimeout to something else than 0.
+func (app *App) ShutdownWithContext(ctx context.Context) error {
+	if app.hooks != nil {
+		// TODO: check should be defered?
+		app.hooks.executeOnShutdownHooks()
+	}
+
 	app.mutex.Lock()
 	defer app.mutex.Unlock()
 	if app.server == nil {
 		return fmt.Errorf("shutdown: server is not running")
 	}
-	return app.server.Shutdown()
+	return app.server.ShutdownWithContext(ctx)
 }
 
 // Server returns the underlying fasthttp server
@@ -602,9 +907,14 @@ func (app *App) Server() *fasthttp.Server {
 	return app.server
 }
 
+// Hooks returns the hook struct to register hooks.
+func (app *App) Hooks() *Hooks {
+	return app.hooks
+}
+
 // Test is used for internal debugging by passing a *http.Request.
 // Timeout is optional and defaults to 1s, -1 will disable it completely.
-func (app *App) Test(req *http.Request, msTimeout ...int) (resp *http.Response, err error) {
+func (app *App) Test(req *http.Request, msTimeout ...int) (*http.Response, error) {
 	// Set timeout
 	timeout := 1000
 	if len(msTimeout) > 0 {
@@ -619,21 +929,31 @@ func (app *App) Test(req *http.Request, msTimeout ...int) (resp *http.Response, 
 	// Dump raw http request
 	dump, err := httputil.DumpRequest(req, true)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("failed to dump request: %w", err)
 	}
 
 	// Create test connection
 	conn := new(testConn)
 
 	// Write raw http request
-	if _, err = conn.r.Write(dump); err != nil {
-		return nil, err
+	if _, err := conn.r.Write(dump); err != nil {
+		return nil, fmt.Errorf("failed to write: %w", err)
 	}
+	// prepare the server for the start
+	app.startupProcess()
 
 	// Serve conn to server
 	channel := make(chan error)
 	go func() {
+		var returned bool
+		defer func() {
+			if !returned {
+				channel <- fmt.Errorf("runtime.Goexit() called in handler or server panic")
+			}
+		}()
+
 		channel <- app.server.ServeConn(conn)
+		returned = true
 	}()
 
 	// Wait for callback
@@ -650,7 +970,7 @@ func (app *App) Test(req *http.Request, msTimeout ...int) (resp *http.Response, 
 	}
 
 	// Check for errors
-	if err != nil && err != fasthttp.ErrGetOnly {
+	if err != nil && !errors.Is(err, fasthttp.ErrGetOnly) {
 		return nil, err
 	}
 
@@ -658,12 +978,17 @@ func (app *App) Test(req *http.Request, msTimeout ...int) (resp *http.Response, 
 	buffer := bufio.NewReader(&conn.w)
 
 	// Convert raw http response to *http.Response
-	return http.ReadResponse(buffer, req)
+	res, err := http.ReadResponse(buffer, req)
+	if err != nil {
+		return nil, fmt.Errorf("failed to read response: %w", err)
+	}
+
+	return res, nil
 }
 
 type disableLogger struct{}
 
-func (dl *disableLogger) Printf(format string, args ...interface{}) {
+func (*disableLogger) Printf(_ string, _ ...any) {
 	// fmt.Println(fmt.Sprintf(format, args...))
 }
 
@@ -671,10 +996,10 @@ func (app *App) init() *App {
 	// lock application
 	app.mutex.Lock()
 
-	// Only load templates if an view engine is specified
+	// Only load templates if a view engine is specified
 	if app.config.Views != nil {
 		if err := app.config.Views.Load(); err != nil {
-			fmt.Printf("views: %v\n", err)
+			log.Warnf("failed to load views: %v", err)
 		}
 	}
 
@@ -682,30 +1007,11 @@ func (app *App) init() *App {
 	app.server = &fasthttp.Server{
 		Logger:       &disableLogger{},
 		LogAllErrors: false,
-		ErrorHandler: func(fctx *fasthttp.RequestCtx, err error) {
-			c := app.AcquireCtx(fctx)
-			if _, ok := err.(*fasthttp.ErrSmallBuffer); ok {
-				err = ErrRequestHeaderFieldsTooLarge
-			} else if netErr, ok := err.(*net.OpError); ok && netErr.Timeout() {
-				err = ErrRequestTimeout
-			} else if err == fasthttp.ErrBodyTooLarge {
-				err = ErrRequestEntityTooLarge
-			} else if err == fasthttp.ErrGetOnly {
-				err = ErrMethodNotAllowed
-			} else if strings.Contains(err.Error(), "timeout") {
-				err = ErrRequestTimeout
-			} else {
-				err = ErrBadRequest
-			}
-			if catch := app.config.ErrorHandler(c, err); catch != nil {
-				_ = c.SendStatus(StatusInternalServerError)
-			}
-			app.ReleaseCtx(c)
-		},
+		ErrorHandler: app.serverErrorHandler,
 	}
 
 	// fasthttp server settings
-	app.server.Handler = app.handler
+	app.server.Handler = app.requestHandler
 	app.server.Name = app.config.ServerHeader
 	app.server.Concurrency = app.config.Concurrency
 	app.server.NoDefaultDate = app.config.DisableDefaultDate
@@ -721,197 +1027,103 @@ func (app *App) init() *App {
 	app.server.WriteBufferSize = app.config.WriteBufferSize
 	app.server.GetOnly = app.config.GETOnly
 	app.server.ReduceMemoryUsage = app.config.ReduceMemoryUsage
+	app.server.StreamRequestBody = app.config.StreamRequestBody
+	app.server.DisablePreParseMultipartForm = app.config.DisablePreParseMultipartForm
 
 	// unlock application
 	app.mutex.Unlock()
 	return app
 }
 
-func (app *App) startupMessage(addr string, tls bool, pids string) {
-	// ignore child processes
-	if IsChild() {
-		return
-	}
-
-	var logo string
-	logo += "%s"
-	logo += " ┌───────────────────────────────────────────────────┐\n"
-	logo += " │ %s │\n"
-	logo += " │ %s │\n"
-	logo += " │                                                   │\n"
-	logo += " │ Handlers %s  Threads %s │\n"
-	logo += " │ Prefork .%s  PID ....%s │\n"
-	logo += " └───────────────────────────────────────────────────┘"
-	logo += "%s"
-
-	const (
-		cBlack = "\u001b[90m"
-		// cRed   = "\u001b[91m"
-		cCyan = "\u001b[96m"
-		// cGreen = "\u001b[92m"
-		// cYellow  = "\u001b[93m"
-		// cBlue    = "\u001b[94m"
-		// cMagenta = "\u001b[95m"
-		// cWhite   = "\u001b[97m"
-		cReset = "\u001b[0m"
+// ErrorHandler is the application's method in charge of finding the
+// appropriate handler for the given request. It searches any mounted
+// sub fibers by their prefixes and if it finds a match, it uses that
+// error handler. Otherwise it uses the configured error handler for
+// the app, which if not set is the DefaultErrorHandler.
+func (app *App) ErrorHandler(ctx Ctx, err error) error {
+	var (
+		mountedErrHandler  ErrorHandler
+		mountedPrefixParts int
 	)
 
-	value := func(s string, width int) string {
-		pad := width - len(s)
-		str := ""
-		for i := 0; i < pad; i++ {
-			str += "."
+	for prefix, subApp := range app.mountFields.appList {
+		if prefix != "" && strings.HasPrefix(ctx.Path(), prefix) {
+			parts := len(strings.Split(prefix, "/"))
+			if mountedPrefixParts <= parts {
+				if subApp.configured.ErrorHandler != nil {
+					mountedErrHandler = subApp.config.ErrorHandler
+				}
+
+				mountedPrefixParts = parts
+			}
 		}
-		if s == "Disabled" {
-			str += " " + s
-		} else {
-			str += fmt.Sprintf(" %s%s%s", cCyan, s, cBlack)
-		}
-		return str
 	}
 
-	center := func(s string, width int) string {
-		pad := strconv.Itoa((width - len(s)) / 2)
-		str := fmt.Sprintf("%"+pad+"s", " ")
-		str += s
-		str += fmt.Sprintf("%"+pad+"s", " ")
-		if len(str) < width {
-			str += " "
-		}
-		return str
+	if mountedErrHandler != nil {
+		return mountedErrHandler(ctx, err)
 	}
 
-	centerValue := func(s string, width int) string {
-		pad := strconv.Itoa((width - len(s)) / 2)
-		str := fmt.Sprintf("%"+pad+"s", " ")
-		str += fmt.Sprintf("%s%s%s", cCyan, s, cBlack)
-		str += fmt.Sprintf("%"+pad+"s", " ")
-		if len(str)-10 < width {
-			str += " "
-		}
-		return str
-	}
+	return app.config.ErrorHandler(ctx, err)
+}
 
-	pad := func(s string, width int) (str string) {
-		toAdd := width - len(s)
-		str += s
-		for i := 0; i < toAdd; i++ {
-			str += " "
-		}
-		return
+// serverErrorHandler is a wrapper around the application's error handler method
+// user for the fasthttp server configuration. It maps a set of fasthttp errors to fiber
+// errors before calling the application's error handler method.
+func (app *App) serverErrorHandler(fctx *fasthttp.RequestCtx, err error) {
+	// Acquire Ctx with fasthttp request from pool
+	c, ok := app.AcquireCtx().(*DefaultCtx)
+	if !ok {
+		panic(fmt.Errorf("failed to type-assert to *DefaultCtx"))
 	}
+	c.Reset(fctx)
 
-	host, port := parseAddr(addr)
-	if host == "" || host == "0.0.0.0" {
-		host = "127.0.0.1"
-	}
-	addr = "http://" + host + ":" + port
-	if tls {
-		addr = "https://" + host + ":" + port
-	}
+	defer app.ReleaseCtx(c)
 
-	isPrefork := "Disabled"
-	if app.config.Prefork {
-		isPrefork = "Enabled"
-	}
-
-	mainLogo := fmt.Sprintf(logo,
-		cBlack,
-		centerValue(" Fiber v"+Version, 49),
-		center(addr, 49),
-		value(strconv.Itoa(app.handlerCount), 14), value(strconv.Itoa(runtime.GOMAXPROCS(0)), 14),
-		value(isPrefork, 14), value(strconv.Itoa(os.Getpid()), 14),
-		cReset,
+	var (
+		errNetOP *net.OpError
+		netErr   net.Error
 	)
 
-	var childPidsLogo string
-	if app.config.Prefork {
-		var childPidsTemplate string
-		childPidsTemplate += "%s"
-		childPidsTemplate += " ┌───────────────────────────────────────────────────┐\n%s"
-		childPidsTemplate += " └───────────────────────────────────────────────────┘"
-		childPidsTemplate += "%s"
-
-		newLine := " │ %s%s%s │"
-
-		// Turn the `pids` variable (in the form ",a,b,c,d,e,f,etc") into a slice of PIDs
-		var pidSlice []string
-		for _, v := range strings.Split(pids, ",") {
-			if v != "" {
-				pidSlice = append(pidSlice, v)
-			}
-		}
-
-		var lines []string
-		thisLine := "Child PIDs ... "
-		var itemsOnThisLine []string
-
-		addLine := func() {
-			lines = append(lines,
-				fmt.Sprintf(
-					newLine,
-					cBlack,
-					thisLine+cCyan+pad(strings.Join(itemsOnThisLine, ", "), 49-len(thisLine)),
-					cBlack,
-				),
-			)
-		}
-
-		for _, pid := range pidSlice {
-			if len(thisLine+strings.Join(append(itemsOnThisLine, pid), ", ")) > 49 {
-				addLine()
-				thisLine = ""
-				itemsOnThisLine = []string{pid}
-			} else {
-				itemsOnThisLine = append(itemsOnThisLine, pid)
-			}
-		}
-
-		// Add left over items to their own line
-		if len(itemsOnThisLine) != 0 {
-			addLine()
-		}
-
-		// Form logo
-		childPidsLogo = fmt.Sprintf(childPidsTemplate,
-			cBlack,
-			strings.Join(lines, "\n")+"\n",
-			cReset,
-		)
+	switch {
+	case errors.As(err, new(*fasthttp.ErrSmallBuffer)):
+		err = ErrRequestHeaderFieldsTooLarge
+	case errors.As(err, &errNetOP) && errNetOP.Timeout():
+		err = ErrRequestTimeout
+	case errors.As(err, &netErr):
+		err = ErrBadGateway
+	case errors.Is(err, fasthttp.ErrBodyTooLarge):
+		err = ErrRequestEntityTooLarge
+	case errors.Is(err, fasthttp.ErrGetOnly):
+		err = ErrMethodNotAllowed
+	case strings.Contains(err.Error(), "timeout"):
+		err = ErrRequestTimeout
+	default:
+		err = NewError(StatusBadRequest, err.Error())
 	}
 
-	// Combine both the child PID logo and the main Fiber logo
-
-	// Pad the shorter logo to the length of the longer one
-	splitMainLogo := strings.Split(mainLogo, "\n")
-	splitChildPidsLogo := strings.Split(childPidsLogo, "\n")
-
-	mainLen := len(splitMainLogo)
-	childLen := len(splitChildPidsLogo)
-
-	if mainLen > childLen {
-		diff := mainLen - childLen
-		for i := 0; i < diff; i++ {
-			splitChildPidsLogo = append(splitChildPidsLogo, "")
-		}
-	} else {
-		diff := childLen - mainLen
-		for i := 0; i < diff; i++ {
-			splitMainLogo = append(splitMainLogo, "")
-		}
+	if catch := app.ErrorHandler(c, err); catch != nil {
+		log.Errorf("serverErrorHandler: failed to call ErrorHandler: %v", catch)
+		_ = c.SendStatus(StatusInternalServerError) //nolint:errcheck // It is fine to ignore the error here
+		return
 	}
+}
 
-	// Combine the two logos, line by line
-	output := "\n"
-	for i := range splitMainLogo {
-		output += cBlack + splitMainLogo[i] + " " + splitChildPidsLogo[i] + "\n"
+// startupProcess Is the method which executes all the necessary processes just before the start of the server.
+func (app *App) startupProcess() *App {
+	app.mutex.Lock()
+	defer app.mutex.Unlock()
+
+	app.mountStartupProcess()
+
+	// build route tree stack
+	app.buildTree()
+
+	return app
+}
+
+// Run onListen hooks. If they return an error, panic.
+func (app *App) runOnListenHooks(listenData ListenData) {
+	if err := app.hooks.executeOnListenHooks(listenData); err != nil {
+		panic(err)
 	}
-
-	out := colorable.NewColorableStdout()
-	if os.Getenv("TERM") == "dumb" || (!isatty.IsTerminal(os.Stdout.Fd()) && !isatty.IsCygwinTerminal(os.Stdout.Fd())) {
-		out = colorable.NewNonColorable(os.Stdout)
-	}
-
-	fmt.Fprintln(out, output)
-
 }
