@@ -1,51 +1,54 @@
+// Package memory Is a slight copy of the memory storage, but far from the storage interface it can not only work with bytes
+// but directly store any kind of data without having to encode it each time, which gives a huge speed advantage
 package memory
 
 import (
 	"sync"
-	"sync/atomic"
 	"time"
+
+	"github.com/gofiber/utils/v2"
 )
 
 type Storage struct {
 	sync.RWMutex
 	data map[string]item // data
-	ts   uint64          // timestamp
 }
 
 type item struct {
-	v interface{} // val
-	e uint64      // exp
+	// max value is 4294967295 -> Sun Feb 07 2106 06:28:15 GMT+0000
+	e uint32 // exp
+	v any    // val
 }
 
 func New() *Storage {
 	store := &Storage{
 		data: make(map[string]item),
-		ts:   uint64(time.Now().Unix()),
 	}
-	go store.gc(10 * time.Millisecond)
-	go store.updater(1 * time.Second)
+	utils.StartTimeStampUpdater()
+	go store.gc(1 * time.Second)
 	return store
 }
 
 // Get value by key
-func (s *Storage) Get(key string) interface{} {
+func (s *Storage) Get(key string) any {
 	s.RLock()
 	v, ok := s.data[key]
 	s.RUnlock()
-	if !ok || v.e != 0 && v.e <= atomic.LoadUint64(&s.ts) {
+	if !ok || v.e != 0 && v.e <= utils.Timestamp() {
 		return nil
 	}
 	return v.v
 }
 
 // Set key with value
-func (s *Storage) Set(key string, val interface{}, ttl time.Duration) {
-	var exp uint64
+func (s *Storage) Set(key string, val any, ttl time.Duration) {
+	var exp uint32
 	if ttl > 0 {
-		exp = uint64(ttl.Seconds()) + atomic.LoadUint64(&s.ts)
+		exp = uint32(ttl.Seconds()) + utils.Timestamp()
 	}
+	i := item{exp, val}
 	s.Lock()
-	s.data[key] = item{val, exp}
+	s.data[key] = i
 	s.Unlock()
 }
 
@@ -58,32 +61,35 @@ func (s *Storage) Delete(key string) {
 
 // Reset all keys
 func (s *Storage) Reset() {
+	nd := make(map[string]item)
 	s.Lock()
-	s.data = make(map[string]item)
+	s.data = nd
 	s.Unlock()
 }
 
-func (s *Storage) updater(sleep time.Duration) {
-	for {
-		time.Sleep(sleep)
-		atomic.StoreUint64(&s.ts, uint64(time.Now().Unix()))
-	}
-}
 func (s *Storage) gc(sleep time.Duration) {
-	expired := []string{}
-	for {
-		time.Sleep(sleep)
+	ticker := time.NewTicker(sleep)
+	defer ticker.Stop()
+	var expired []string
+
+	for range ticker.C {
+		ts := utils.Timestamp()
 		expired = expired[:0]
 		s.RLock()
 		for key, v := range s.data {
-			if v.e != 0 && v.e <= atomic.LoadUint64(&s.ts) {
+			if v.e != 0 && v.e <= ts {
 				expired = append(expired, key)
 			}
 		}
 		s.RUnlock()
 		s.Lock()
+		// Double-checked locking.
+		// We might have replaced the item in the meantime.
 		for i := range expired {
-			delete(s.data, expired[i])
+			v := s.data[expired[i]]
+			if v.e != 0 && v.e <= ts {
+				delete(s.data, expired[i])
+			}
 		}
 		s.Unlock()
 	}
