@@ -1,32 +1,30 @@
 package basicauth
 
 import (
+	"encoding/base64"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http/httptest"
 	"testing"
 
-	b64 "encoding/base64"
-
-	"github.com/gofiber/fiber/v2"
-	"github.com/gofiber/fiber/v2/utils"
+	"github.com/gofiber/fiber/v3"
+	"github.com/stretchr/testify/require"
 	"github.com/valyala/fasthttp"
 )
 
 // go test -run Test_BasicAuth_Next
 func Test_BasicAuth_Next(t *testing.T) {
 	t.Parallel()
-
 	app := fiber.New()
 	app.Use(New(Config{
-		Next: func(_ *fiber.Ctx) bool {
+		Next: func(_ fiber.Ctx) bool {
 			return true
 		},
 	}))
 
-	resp, err := app.Test(httptest.NewRequest("GET", "/", nil))
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, fiber.StatusNotFound, resp.StatusCode)
+	resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/", nil))
+	require.NoError(t, err)
+	require.Equal(t, fiber.StatusNotFound, resp.StatusCode)
 }
 
 func Test_Middleware_BasicAuth(t *testing.T) {
@@ -40,9 +38,9 @@ func Test_Middleware_BasicAuth(t *testing.T) {
 		},
 	}))
 
-	app.Get("/testauth", func(c *fiber.Ctx) error {
-		username := c.Locals("username").(string)
-		password := c.Locals("password").(string)
+	app.Get("/testauth", func(c fiber.Ctx) error {
+		username := UsernameFromContext(c)
+		password := PasswordFromContext(c)
 
 		return c.SendString(username + password)
 	})
@@ -75,20 +73,20 @@ func Test_Middleware_BasicAuth(t *testing.T) {
 
 	for _, tt := range tests {
 		// Base64 encode credentials for http auth header
-		creds := b64.StdEncoding.EncodeToString([]byte(fmt.Sprintf("%s:%s", tt.username, tt.password)))
+		creds := base64.StdEncoding.EncodeToString([]byte(fmt.Sprintf("%s:%s", tt.username, tt.password)))
 
-		req := httptest.NewRequest("GET", "/testauth", nil)
+		req := httptest.NewRequest(fiber.MethodGet, "/testauth", nil)
 		req.Header.Add("Authorization", "Basic "+creds)
 		resp, err := app.Test(req)
-		utils.AssertEqual(t, nil, err)
+		require.NoError(t, err)
 
-		body, err := ioutil.ReadAll(resp.Body)
+		body, err := io.ReadAll(resp.Body)
 
-		utils.AssertEqual(t, nil, err)
-		utils.AssertEqual(t, tt.statusCode, resp.StatusCode)
+		require.NoError(t, err)
+		require.Equal(t, tt.statusCode, resp.StatusCode)
 
 		if tt.statusCode == 200 {
-			utils.AssertEqual(t, fmt.Sprintf("%s%s", tt.username, tt.password), string(body))
+			require.Equal(t, fmt.Sprintf("%s%s", tt.username, tt.password), string(body))
 		}
 	}
 }
@@ -102,14 +100,14 @@ func Benchmark_Middleware_BasicAuth(b *testing.B) {
 			"john": "doe",
 		},
 	}))
-	app.Get("/", func(c *fiber.Ctx) error {
+	app.Get("/", func(c fiber.Ctx) error {
 		return c.SendStatus(fiber.StatusTeapot)
 	})
 
 	h := app.Handler()
 
 	fctx := &fasthttp.RequestCtx{}
-	fctx.Request.Header.SetMethod("GET")
+	fctx.Request.Header.SetMethod(fiber.MethodGet)
 	fctx.Request.SetRequestURI("/")
 	fctx.Request.Header.Set(fiber.HeaderAuthorization, "basic am9objpkb2U=") // john:doe
 
@@ -120,5 +118,35 @@ func Benchmark_Middleware_BasicAuth(b *testing.B) {
 		h(fctx)
 	}
 
-	utils.AssertEqual(b, fiber.StatusTeapot, fctx.Response.Header.StatusCode())
+	require.Equal(b, fiber.StatusTeapot, fctx.Response.Header.StatusCode())
+}
+
+// go test -v -run=^$ -bench=Benchmark_Middleware_BasicAuth -benchmem -count=4
+func Benchmark_Middleware_BasicAuth_Upper(b *testing.B) {
+	app := fiber.New()
+
+	app.Use(New(Config{
+		Users: map[string]string{
+			"john": "doe",
+		},
+	}))
+	app.Get("/", func(c fiber.Ctx) error {
+		return c.SendStatus(fiber.StatusTeapot)
+	})
+
+	h := app.Handler()
+
+	fctx := &fasthttp.RequestCtx{}
+	fctx.Request.Header.SetMethod(fiber.MethodGet)
+	fctx.Request.SetRequestURI("/")
+	fctx.Request.Header.Set(fiber.HeaderAuthorization, "Basic am9objpkb2U=") // john:doe
+
+	b.ReportAllocs()
+	b.ResetTimer()
+
+	for n := 0; n < b.N; n++ {
+		h(fctx)
+	}
+
+	require.Equal(b, fiber.StatusTeapot, fctx.Response.Header.StatusCode())
 }
