@@ -4,19 +4,21 @@ import (
 	"sync"
 	"time"
 
-	"github.com/gofiber/fiber/v2"
-	"github.com/gofiber/fiber/v2/internal/memory"
+	"github.com/gofiber/fiber/v3"
+	"github.com/gofiber/fiber/v3/internal/memory"
 )
 
 // go:generate msgp
 // msgp -file="manager.go" -o="manager_msgp.go" -tests=false -unexported
-// don't forget to replace the msgp import path to:
-// "github.com/gofiber/fiber/v2/internal/msgp"
 type item struct {
-	body   []byte
-	ctype  []byte
-	status int
-	exp    uint64
+	body      []byte
+	ctype     []byte
+	cencoding []byte
+	status    int
+	exp       uint64
+	headers   map[string][]byte
+	// used for finding the item in an indexed heap
+	heapidx int
 }
 
 //msgp:ignore manager
@@ -30,7 +32,7 @@ func newManager(storage fiber.Storage) *manager {
 	// Create new storage handler
 	manager := &manager{
 		pool: sync.Pool{
-			New: func() interface{} {
+			New: func() any {
 				return new(item)
 			},
 		},
@@ -39,7 +41,7 @@ func newManager(storage fiber.Storage) *manager {
 		// Use provided storage if provided
 		manager.storage = storage
 	} else {
-		// Fallback too memory storage
+		// Fallback to memory storage
 		manager.memory = memory.New()
 	}
 	return manager
@@ -47,7 +49,7 @@ func newManager(storage fiber.Storage) *manager {
 
 // acquire returns an *entry from the sync.Pool
 func (m *manager) acquire() *item {
-	return m.pool.Get().(*item)
+	return m.pool.Get().(*item) //nolint:forcetypeassert // We store nothing else in the pool
 }
 
 // release and reset *entry to sync.Pool
@@ -60,43 +62,52 @@ func (m *manager) release(e *item) {
 	e.ctype = nil
 	e.status = 0
 	e.exp = 0
+	e.headers = nil
 	m.pool.Put(e)
 }
 
 // get data from storage or memory
-func (m *manager) get(key string) (it *item) {
+func (m *manager) get(key string) *item {
+	var it *item
 	if m.storage != nil {
 		it = m.acquire()
-		if raw, _ := m.storage.Get(key); raw != nil {
+		raw, err := m.storage.Get(key)
+		if err != nil {
+			return it
+		}
+		if raw != nil {
 			if _, err := it.UnmarshalMsg(raw); err != nil {
-				return
+				return it
 			}
 		}
-		return
+		return it
 	}
-	if it, _ = m.memory.Get(key).(*item); it == nil {
+	if it, _ = m.memory.Get(key).(*item); it == nil { //nolint:errcheck // We store nothing else in the pool
 		it = m.acquire()
+		return it
 	}
-	return
-
+	return it
 }
 
 // get raw data from storage or memory
-func (m *manager) getRaw(key string) (raw []byte) {
+func (m *manager) getRaw(key string) []byte {
+	var raw []byte
 	if m.storage != nil {
-		raw, _ = m.storage.Get(key)
+		raw, _ = m.storage.Get(key) //nolint:errcheck // TODO: Handle error here
 	} else {
-		raw, _ = m.memory.Get(key).([]byte)
+		raw, _ = m.memory.Get(key).([]byte) //nolint:errcheck // TODO: Handle error here
 	}
-	return
+	return raw
 }
 
 // set data to storage or memory
 func (m *manager) set(key string, it *item, exp time.Duration) {
 	if m.storage != nil {
 		if raw, err := it.MarshalMsg(nil); err == nil {
-			_ = m.storage.Set(key, raw, exp)
+			_ = m.storage.Set(key, raw, exp) //nolint:errcheck // TODO: Handle error here
 		}
+		// we can release data because it's serialized to database
+		m.release(it)
 	} else {
 		m.memory.Set(key, it, exp)
 	}
@@ -105,16 +116,16 @@ func (m *manager) set(key string, it *item, exp time.Duration) {
 // set data to storage or memory
 func (m *manager) setRaw(key string, raw []byte, exp time.Duration) {
 	if m.storage != nil {
-		_ = m.storage.Set(key, raw, exp)
+		_ = m.storage.Set(key, raw, exp) //nolint:errcheck // TODO: Handle error here
 	} else {
 		m.memory.Set(key, raw, exp)
 	}
 }
 
 // delete data from storage or memory
-func (m *manager) delete(key string) {
+func (m *manager) del(key string) {
 	if m.storage != nil {
-		_ = m.storage.Delete(key)
+		_ = m.storage.Delete(key) //nolint:errcheck // TODO: Handle error here
 	} else {
 		m.memory.Delete(key)
 	}
