@@ -6,57 +6,58 @@ import (
 
 // go:generate msgp
 // msgp -file="data.go" -o="data_msgp.go" -tests=false -unexported
-// don't forget to replace the msgp import path to:
-// "github.com/gofiber/fiber/v2/internal/msgp"
 type data struct {
-	sync.RWMutex `gotiny:"-"`
-	d            map[string]interface{} `gotiny:"d"`
+	sync.RWMutex
+	Data map[string]any
 }
 
 var dataPool = sync.Pool{
-	New: func() interface{} {
+	New: func() any {
 		d := new(data)
-		d.d = make(map[string]interface{})
+		d.Data = make(map[string]any)
 		return d
 	},
 }
 
 func acquireData() *data {
-	return dataPool.Get().(*data)
-}
-
-func releaseData(d *data) {
-	d.Reset()
-	dataPool.Put(d)
+	return dataPool.Get().(*data) //nolint:forcetypeassert // We store nothing else in the pool
 }
 
 func (d *data) Reset() {
 	d.Lock()
-	for key := range d.d {
-		delete(d.d, key)
-	}
+	d.Data = make(map[string]any)
 	d.Unlock()
 }
 
-func (d *data) Get(key string) interface{} {
+func (d *data) Get(key string) any {
 	d.RLock()
-	v := d.d[key]
+	v := d.Data[key]
 	d.RUnlock()
 	return v
 }
 
-func (d *data) Set(key string, value interface{}) {
+func (d *data) Set(key string, value any) {
 	d.Lock()
-	d.d[key] = value
+	d.Data[key] = value
 	d.Unlock()
 }
 
 func (d *data) Delete(key string) {
 	d.Lock()
-	delete(d.d, key)
+	delete(d.Data, key)
 	d.Unlock()
 }
 
+func (d *data) Keys() []string {
+	d.Lock()
+	keys := make([]string, 0, len(d.Data))
+	for k := range d.Data {
+		keys = append(keys, k)
+	}
+	d.Unlock()
+	return keys
+}
+
 func (d *data) Len() int {
-	return len(d.d)
+	return len(d.Data)
 }
