@@ -1,34 +1,40 @@
 package session
 
 import (
+	"bytes"
+	"encoding/gob"
+	"fmt"
 	"sync"
 	"time"
 
-	"github.com/gofiber/fiber/v2"
-	"github.com/gofiber/fiber/v2/internal/gotiny"
-	"github.com/gofiber/fiber/v2/utils"
+	"github.com/gofiber/fiber/v3"
+	"github.com/gofiber/utils/v2"
 	"github.com/valyala/fasthttp"
 )
 
 type Session struct {
-	// sync.RWMutex
-	id     string     // session id
-	fresh  bool       // if new session
-	ctx    *fiber.Ctx // fiber context
-	config *Store     // store configuration
-	data   *data      // key value data
+	id         string        // session id
+	fresh      bool          // if new session
+	ctx        fiber.Ctx     // fiber context
+	config     *Store        // store configuration
+	data       *data         // key value data
+	byteBuffer *bytes.Buffer // byte buffer for the en- and decode
+	exp        time.Duration // expiration of this session
 }
 
 var sessionPool = sync.Pool{
-	New: func() interface{} {
+	New: func() any {
 		return new(Session)
 	},
 }
 
 func acquireSession() *Session {
-	s := sessionPool.Get().(*Session)
+	s := sessionPool.Get().(*Session) //nolint:forcetypeassert,errcheck // We store nothing else in the pool
 	if s.data == nil {
 		s.data = acquireData()
+	}
+	if s.byteBuffer == nil {
+		s.byteBuffer = new(bytes.Buffer)
 	}
 	s.fresh = true
 	return s
@@ -36,10 +42,14 @@ func acquireSession() *Session {
 
 func releaseSession(s *Session) {
 	s.id = ""
+	s.exp = 0
 	s.ctx = nil
 	s.config = nil
 	if s.data != nil {
 		s.data.Reset()
+	}
+	if s.byteBuffer != nil {
+		s.byteBuffer.Reset()
 	}
 	sessionPool.Put(s)
 }
@@ -55,7 +65,7 @@ func (s *Session) ID() string {
 }
 
 // Get will return the value
-func (s *Session) Get(key string) interface{} {
+func (s *Session) Get(key string) any {
 	// Better safe than sorry
 	if s.data == nil {
 		return nil
@@ -64,7 +74,7 @@ func (s *Session) Get(key string) interface{} {
 }
 
 // Set will update or create a new key value
-func (s *Session) Set(key string, val interface{}) {
+func (s *Session) Set(key string, val any) {
 	// Better safe than sorry
 	if s.data == nil {
 		return
@@ -96,51 +106,91 @@ func (s *Session) Destroy() error {
 		return err
 	}
 
-	// Expire cookie
-	s.delCookie()
+	// Expire session
+	s.delSession()
 	return nil
 }
 
 // Regenerate generates a new session id and delete the old one from Storage
 func (s *Session) Regenerate() error {
+	// Delete old id from storage
+	if err := s.config.Storage.Delete(s.id); err != nil {
+		return err
+	}
+
+	// Generate a new session, and set session.fresh to true
+	s.refresh()
+
+	return nil
+}
+
+// Reset generates a new session id, deletes the old one from storage, and resets the associated data
+func (s *Session) Reset() error {
+	// Reset local data
+	if s.data != nil {
+		s.data.Reset()
+	}
+	// Reset byte buffer
+	if s.byteBuffer != nil {
+		s.byteBuffer.Reset()
+	}
+	// Reset expiration
+	s.exp = 0
 
 	// Delete old id from storage
 	if err := s.config.Storage.Delete(s.id); err != nil {
 		return err
 	}
 
-	// Create new ID
-	s.id = s.config.KeyGenerator()
+	// Expire session
+	s.delSession()
+
+	// Generate a new session, and set session.fresh to true
+	s.refresh()
 
 	return nil
 }
 
+// refresh generates a new session, and set session.fresh to be true
+func (s *Session) refresh() {
+	// Create a new id
+	s.id = s.config.KeyGenerator()
+
+	// We assign a new id to the session, so the session must be fresh
+	s.fresh = true
+}
+
 // Save will update the storage and client cookie
 func (s *Session) Save() error {
-
 	// Better safe than sorry
 	if s.data == nil {
 		return nil
 	}
 
-	// Don't save to Storage if no data is available
-	if s.data.Len() <= 0 {
-		return nil
+	// Check if session has your own expiration, otherwise use default value
+	if s.exp <= 0 {
+		s.exp = s.config.Expiration
 	}
+
+	// Update client cookie
+	s.setSession()
 
 	// Convert data to bytes
-	// s.RLock()
-	data := gotiny.Marshal(&s.data)
-	// s.RUnlock()
-
-	// pass raw bytes with session id to provider
-	if err := s.config.Storage.Set(s.id, data, s.config.Expiration); err != nil {
-		return err
+	mux.Lock()
+	defer mux.Unlock()
+	encCache := gob.NewEncoder(s.byteBuffer)
+	err := encCache.Encode(&s.data.Data)
+	if err != nil {
+		return fmt.Errorf("failed to encode data: %w", err)
 	}
 
-	// Create cookie with the session ID if fresh
-	if s.fresh {
-		s.setCookie()
+	// copy the data in buffer
+	encodedBytes := make([]byte, s.byteBuffer.Len())
+	copy(encodedBytes, s.byteBuffer.Bytes())
+
+	// pass copied bytes with session id to provider
+	if err := s.config.Storage.Set(s.id, encodedBytes, s.exp); err != nil {
+		return err
 	}
 
 	// Release session
@@ -150,52 +200,78 @@ func (s *Session) Save() error {
 	return nil
 }
 
-func (s *Session) setCookie() {
-	fcookie := fasthttp.AcquireCookie()
-	fcookie.SetKey(s.config.CookieName)
-	fcookie.SetValue(s.id)
-	fcookie.SetPath(s.config.CookiePath)
-	fcookie.SetDomain(s.config.CookieDomain)
-	fcookie.SetMaxAge(int(s.config.Expiration.Seconds()))
-	fcookie.SetExpire(time.Now().Add(s.config.Expiration))
-	fcookie.SetSecure(s.config.CookieSecure)
-	fcookie.SetHTTPOnly(s.config.CookieHTTPOnly)
-
-	switch utils.ToLower(s.config.CookieSameSite) {
-	case "strict":
-		fcookie.SetSameSite(fasthttp.CookieSameSiteStrictMode)
-	case "none":
-		fcookie.SetSameSite(fasthttp.CookieSameSiteNoneMode)
-	default:
-		fcookie.SetSameSite(fasthttp.CookieSameSiteLaxMode)
+// Keys will retrieve all keys in current session
+func (s *Session) Keys() []string {
+	if s.data == nil {
+		return []string{}
 	}
-
-	s.ctx.Response().Header.SetCookie(fcookie)
-	fasthttp.ReleaseCookie(fcookie)
+	return s.data.Keys()
 }
 
-func (s *Session) delCookie() {
-	s.ctx.Request().Header.DelCookie(s.config.CookieName)
-	s.ctx.Response().Header.DelCookie(s.config.CookieName)
+// SetExpiry sets a specific expiration for this session
+func (s *Session) SetExpiry(exp time.Duration) {
+	s.exp = exp
+}
 
-	fcookie := fasthttp.AcquireCookie()
-	fcookie.SetKey(s.config.CookieName)
-	fcookie.SetPath(s.config.CookiePath)
-	fcookie.SetDomain(s.config.CookieDomain)
-	fcookie.SetMaxAge(-1)
-	fcookie.SetExpire(time.Now().Add(-1 * time.Minute))
-	fcookie.SetSecure(s.config.CookieSecure)
-	fcookie.SetHTTPOnly(s.config.CookieHTTPOnly)
+func (s *Session) setSession() {
+	if s.config.source == SourceHeader {
+		s.ctx.Request().Header.SetBytesV(s.config.sessionName, []byte(s.id))
+		s.ctx.Response().Header.SetBytesV(s.config.sessionName, []byte(s.id))
+	} else {
+		fcookie := fasthttp.AcquireCookie()
+		fcookie.SetKey(s.config.sessionName)
+		fcookie.SetValue(s.id)
+		fcookie.SetPath(s.config.CookiePath)
+		fcookie.SetDomain(s.config.CookieDomain)
+		// Cookies are also session cookies if they do not specify the Expires or Max-Age attribute.
+		// refer: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie
+		if !s.config.CookieSessionOnly {
+			fcookie.SetMaxAge(int(s.exp.Seconds()))
+			fcookie.SetExpire(time.Now().Add(s.exp))
+		}
+		fcookie.SetSecure(s.config.CookieSecure)
+		fcookie.SetHTTPOnly(s.config.CookieHTTPOnly)
 
-	switch utils.ToLower(s.config.CookieSameSite) {
-	case "strict":
-		fcookie.SetSameSite(fasthttp.CookieSameSiteStrictMode)
-	case "none":
-		fcookie.SetSameSite(fasthttp.CookieSameSiteNoneMode)
-	default:
-		fcookie.SetSameSite(fasthttp.CookieSameSiteLaxMode)
+		switch utils.ToLower(s.config.CookieSameSite) {
+		case "strict":
+			fcookie.SetSameSite(fasthttp.CookieSameSiteStrictMode)
+		case "none":
+			fcookie.SetSameSite(fasthttp.CookieSameSiteNoneMode)
+		default:
+			fcookie.SetSameSite(fasthttp.CookieSameSiteLaxMode)
+		}
+		s.ctx.Response().Header.SetCookie(fcookie)
+		fasthttp.ReleaseCookie(fcookie)
 	}
+}
 
-	s.ctx.Response().Header.SetCookie(fcookie)
-	fasthttp.ReleaseCookie(fcookie)
+func (s *Session) delSession() {
+	if s.config.source == SourceHeader {
+		s.ctx.Request().Header.Del(s.config.sessionName)
+		s.ctx.Response().Header.Del(s.config.sessionName)
+	} else {
+		s.ctx.Request().Header.DelCookie(s.config.sessionName)
+		s.ctx.Response().Header.DelCookie(s.config.sessionName)
+
+		fcookie := fasthttp.AcquireCookie()
+		fcookie.SetKey(s.config.sessionName)
+		fcookie.SetPath(s.config.CookiePath)
+		fcookie.SetDomain(s.config.CookieDomain)
+		fcookie.SetMaxAge(-1)
+		fcookie.SetExpire(time.Now().Add(-1 * time.Minute))
+		fcookie.SetSecure(s.config.CookieSecure)
+		fcookie.SetHTTPOnly(s.config.CookieHTTPOnly)
+
+		switch utils.ToLower(s.config.CookieSameSite) {
+		case "strict":
+			fcookie.SetSameSite(fasthttp.CookieSameSiteStrictMode)
+		case "none":
+			fcookie.SetSameSite(fasthttp.CookieSameSiteNoneMode)
+		default:
+			fcookie.SetSameSite(fasthttp.CookieSameSiteLaxMode)
+		}
+
+		s.ctx.Response().Header.SetCookie(fcookie)
+		fasthttp.ReleaseCookie(fcookie)
+	}
 }
