@@ -1,9 +1,12 @@
+// Package memory Is a copy of the storage memory from the external storage packet as a purpose to test the behavior
+// in the unittests when using a storages from these packets
 package memory
 
 import (
-	"errors"
 	"sync"
 	"time"
+
+	"github.com/gofiber/utils/v2"
 )
 
 // Storage interface that is implemented by storage providers
@@ -14,24 +17,26 @@ type Storage struct {
 	done       chan struct{}
 }
 
-// Common storage errors
-var ErrNotExist = errors.New("key does not exist")
-
 type entry struct {
-	data   []byte
-	expiry int64
+	data []byte
+	// max value is 4294967295 -> Sun Feb 07 2106 06:28:15 GMT+0000
+	expiry uint32
 }
 
 // New creates a new memory storage
-func New() *Storage {
+func New(config ...Config) *Storage {
+	// Set default config
+	cfg := configDefault(config...)
+
 	// Create storage
 	store := &Storage{
 		db:         make(map[string]entry),
-		gcInterval: 10 * time.Second,
+		gcInterval: cfg.GCInterval,
 		done:       make(chan struct{}),
 	}
 
 	// Start garbage collector
+	utils.StartTimeStampUpdater()
 	go store.gc()
 
 	return store
@@ -40,19 +45,18 @@ func New() *Storage {
 // Get value by key
 func (s *Storage) Get(key string) ([]byte, error) {
 	if len(key) <= 0 {
-		return nil, ErrNotExist
+		return nil, nil
 	}
 	s.mux.RLock()
 	v, ok := s.db[key]
 	s.mux.RUnlock()
-	if !ok || v.expiry != 0 && v.expiry <= time.Now().Unix() {
-		return nil, ErrNotExist
+	if !ok || v.expiry != 0 && v.expiry <= utils.Timestamp() {
+		return nil, nil
 	}
 
 	return v.data, nil
 }
 
-// Set key with value
 // Set key with value
 func (s *Storage) Set(key string, val []byte, exp time.Duration) error {
 	// Ain't Nobody Got Time For That
@@ -60,13 +64,14 @@ func (s *Storage) Set(key string, val []byte, exp time.Duration) error {
 		return nil
 	}
 
-	var expire int64
+	var expire uint32
 	if exp != 0 {
-		expire = time.Now().Add(exp).Unix()
+		expire = uint32(exp.Seconds()) + utils.Timestamp()
 	}
 
+	e := entry{val, expire}
 	s.mux.Lock()
-	s.db[key] = entry{val, expire}
+	s.db[key] = e
 	s.mux.Unlock()
 	return nil
 }
@@ -85,8 +90,9 @@ func (s *Storage) Delete(key string) error {
 
 // Reset all keys
 func (s *Storage) Reset() error {
+	ndb := make(map[string]entry)
 	s.mux.Lock()
-	s.db = make(map[string]entry)
+	s.db = ndb
 	s.mux.Unlock()
 	return nil
 }
@@ -100,20 +106,39 @@ func (s *Storage) Close() error {
 func (s *Storage) gc() {
 	ticker := time.NewTicker(s.gcInterval)
 	defer ticker.Stop()
+	var expired []string
 
 	for {
 		select {
 		case <-s.done:
 			return
-		case t := <-ticker.C:
-			now := t.Unix()
-			s.mux.Lock()
+		case <-ticker.C:
+			ts := utils.Timestamp()
+			expired = expired[:0]
+			s.mux.RLock()
 			for id, v := range s.db {
-				if v.expiry != 0 && v.expiry < now {
-					delete(s.db, id)
+				if v.expiry != 0 && v.expiry <= ts {
+					expired = append(expired, id)
+				}
+			}
+			s.mux.RUnlock()
+			s.mux.Lock()
+			// Double-checked locking.
+			// We might have replaced the item in the meantime.
+			for i := range expired {
+				v := s.db[expired[i]]
+				if v.expiry != 0 && v.expiry <= ts {
+					delete(s.db, expired[i])
 				}
 			}
 			s.mux.Unlock()
 		}
 	}
+}
+
+// Return database client
+func (s *Storage) Conn() map[string]entry {
+	s.mux.RLock()
+	defer s.mux.RUnlock()
+	return s.db
 }
