@@ -1,6 +1,8 @@
 package cors
 
-import "strings"
+import (
+	"strings"
+)
 
 func matchScheme(domain, pattern string) bool {
 	didx := strings.Index(domain, ":")
@@ -20,18 +22,20 @@ func matchSubdomain(domain, pattern string) bool {
 	}
 	domAuth := domain[didx+3:]
 	// to avoid long loop by invalid long domain
-	if len(domAuth) > 253 {
+	const maxDomainLen = 253
+	if len(domAuth) > maxDomainLen {
 		return false
 	}
 	patAuth := pattern[pidx+3:]
 
 	domComp := strings.Split(domAuth, ".")
 	patComp := strings.Split(patAuth, ".")
-	for i := len(domComp)/2 - 1; i >= 0; i-- {
+	const divHalf = 2
+	for i := len(domComp)/divHalf - 1; i >= 0; i-- {
 		opp := len(domComp) - 1 - i
 		domComp[i], domComp[opp] = domComp[opp], domComp[i]
 	}
-	for i := len(patComp)/2 - 1; i >= 0; i-- {
+	for i := len(patComp)/divHalf - 1; i >= 0; i-- {
 		opp := len(patComp) - 1 - i
 		patComp[i], patComp[opp] = patComp[opp], patComp[i]
 	}
