@@ -1,35 +1,44 @@
+//nolint:bodyclose // Much easier to just ignore memory leaks in tests
 package filesystem
 
 import (
+	"context"
 	"net/http"
 	"net/http/httptest"
+	"os"
 	"testing"
 
-	"github.com/gofiber/fiber/v2"
-	"github.com/gofiber/fiber/v2/utils"
+	"github.com/gofiber/fiber/v3"
+	"github.com/stretchr/testify/require"
 )
 
 // go test -run Test_FileSystem
 func Test_FileSystem(t *testing.T) {
+	t.Parallel()
 	app := fiber.New()
 
 	app.Use("/test", New(Config{
-		Root: http.Dir("../../.github/testdata/fs"),
+		Root: os.DirFS("../../.github/testdata/fs"),
 	}))
 
 	app.Use("/dir", New(Config{
-		Root:   http.Dir("../../.github/testdata/fs"),
+		Root:   os.DirFS("../../.github/testdata/fs"),
 		Browse: true,
 	}))
 
-	app.Get("/", func(c *fiber.Ctx) error {
+	app.Get("/", func(c fiber.Ctx) error {
 		return c.SendString("Hello, World!")
 	})
 
 	app.Use("/spatest", New(Config{
-		Root:         http.Dir("../../.github/testdata/fs"),
+		Root:         os.DirFS("../../.github/testdata/fs"),
 		Index:        "index.html",
 		NotFoundFile: "index.html",
+	}))
+
+	app.Use("/prefix", New(Config{
+		Root:       os.DirFS("../../.github/testdata/fs"),
+		PathPrefix: "img",
 	}))
 
 	tests := []struct {
@@ -85,6 +94,12 @@ func Test_FileSystem(t *testing.T) {
 			contentType: "text/html",
 		},
 		{
+			name:        "Should list the directory contents",
+			url:         "/dir/img/",
+			statusCode:  200,
+			contentType: "text/html",
+		},
+		{
 			name:        "Should be returns status 200",
 			url:         "/dir/img/fiber.png",
 			statusCode:  200,
@@ -96,17 +111,25 @@ func Test_FileSystem(t *testing.T) {
 			statusCode:  200,
 			contentType: "text/html",
 		},
+		{
+			name:        "PathPrefix should be applied",
+			url:         "/prefix/fiber.png",
+			statusCode:  200,
+			contentType: "image/png",
+		},
 	}
 
 	for _, tt := range tests {
+		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
-			resp, err := app.Test(httptest.NewRequest("GET", tt.url, nil))
-			utils.AssertEqual(t, nil, err)
-			utils.AssertEqual(t, tt.statusCode, resp.StatusCode)
+			t.Parallel()
+			resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, tt.url, nil))
+			require.NoError(t, err)
+			require.Equal(t, tt.statusCode, resp.StatusCode)
 
 			if tt.contentType != "" {
 				ct := resp.Header.Get("Content-Type")
-				utils.AssertEqual(t, tt.contentType, ct)
+				require.Equal(t, tt.contentType, ct)
 			}
 		})
 	}
@@ -114,76 +137,117 @@ func Test_FileSystem(t *testing.T) {
 
 // go test -run Test_FileSystem_Next
 func Test_FileSystem_Next(t *testing.T) {
+	t.Parallel()
 	app := fiber.New()
 	app.Use(New(Config{
-		Root: http.Dir("../../.github/testdata/fs"),
-		Next: func(_ *fiber.Ctx) bool {
+		Root: os.DirFS("../../.github/testdata/fs"),
+		Next: func(_ fiber.Ctx) bool {
 			return true
 		},
 	}))
 
-	resp, err := app.Test(httptest.NewRequest("GET", "/", nil))
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, fiber.StatusNotFound, resp.StatusCode)
+	resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/", nil))
+	require.NoError(t, err)
+	require.Equal(t, fiber.StatusNotFound, resp.StatusCode)
+}
+
+// go test -run Test_FileSystem_Download
+func Test_FileSystem_Download(t *testing.T) {
+	app := fiber.New()
+	app.Use(New(Config{
+		Root:     os.DirFS("../../.github/testdata/fs"),
+		Download: true,
+	}))
+
+	resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/img/fiber.png", nil))
+	require.NoError(t, err, "app.Test(req)")
+	require.Equal(t, 200, resp.StatusCode, "Status code")
+	require.NotEmpty(t, resp.Header.Get(fiber.HeaderContentLength))
+	require.Equal(t, "image/png", resp.Header.Get(fiber.HeaderContentType))
+	require.Equal(t, "attachment", resp.Header.Get(fiber.HeaderContentDisposition))
 }
 
 func Test_FileSystem_NonGetAndHead(t *testing.T) {
+	t.Parallel()
 	app := fiber.New()
 
 	app.Use("/test", New(Config{
-		Root: http.Dir("../../.github/testdata/fs"),
+		Root: os.DirFS("../../.github/testdata/fs"),
 	}))
 
 	resp, err := app.Test(httptest.NewRequest(fiber.MethodPost, "/test", nil))
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, 404, resp.StatusCode)
+	require.NoError(t, err)
+	require.Equal(t, 404, resp.StatusCode)
 }
 
 func Test_FileSystem_Head(t *testing.T) {
+	t.Parallel()
 	app := fiber.New()
 
 	app.Use("/test", New(Config{
-		Root: http.Dir("../../.github/testdata/fs"),
+		Root: os.DirFS("../../.github/testdata/fs"),
 	}))
 
-	req, _ := http.NewRequest(fiber.MethodHead, "/test", nil)
+	req, err := http.NewRequestWithContext(context.Background(), fiber.MethodHead, "/test", nil)
+	require.NoError(t, err)
 	resp, err := app.Test(req)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, 200, resp.StatusCode)
+	require.NoError(t, err)
+	require.Equal(t, 200, resp.StatusCode)
 }
 
 func Test_FileSystem_NoRoot(t *testing.T) {
+	t.Parallel()
 	defer func() {
-		utils.AssertEqual(t, "filesystem: Root cannot be nil", recover())
+		require.Equal(t, "filesystem: Root cannot be nil", recover())
 	}()
 
 	app := fiber.New()
 	app.Use(New())
-	_, _ = app.Test(httptest.NewRequest(fiber.MethodGet, "/", nil))
+	_, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/", nil))
+	require.NoError(t, err)
 }
 
 func Test_FileSystem_UsingParam(t *testing.T) {
+	t.Parallel()
 	app := fiber.New()
 
-	app.Use("/:path", func(c *fiber.Ctx) error {
-		return SendFile(c, http.Dir("../../.github/testdata/fs"), c.Params("path")+".html")
+	app.Use("/:path", func(c fiber.Ctx) error {
+		return SendFile(c, os.DirFS("../../.github/testdata/fs"), c.Params("path")+".html")
 	})
 
-	req, _ := http.NewRequest(fiber.MethodHead, "/index", nil)
+	req, err := http.NewRequestWithContext(context.Background(), fiber.MethodHead, "/index", nil)
+	require.NoError(t, err)
 	resp, err := app.Test(req)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, 200, resp.StatusCode)
+	require.NoError(t, err)
+	require.Equal(t, 200, resp.StatusCode)
 }
 
 func Test_FileSystem_UsingParam_NonFile(t *testing.T) {
+	t.Parallel()
 	app := fiber.New()
 
-	app.Use("/:path", func(c *fiber.Ctx) error {
-		return SendFile(c, http.Dir("../../.github/testdata/fs"), c.Params("path")+".html")
+	app.Use("/:path", func(c fiber.Ctx) error {
+		return SendFile(c, os.DirFS("../../.github/testdata/fs"), c.Params("path")+".html")
 	})
 
-	req, _ := http.NewRequest(fiber.MethodHead, "/template", nil)
+	req, err := http.NewRequestWithContext(context.Background(), fiber.MethodHead, "/template", nil)
+	require.NoError(t, err)
 	resp, err := app.Test(req)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, 404, resp.StatusCode)
+	require.NoError(t, err)
+	require.Equal(t, 404, resp.StatusCode)
+}
+
+func Test_FileSystem_UsingContentTypeCharset(t *testing.T) {
+	t.Parallel()
+	app := fiber.New()
+	app.Use(New(Config{
+		Root:               os.DirFS("../../.github/testdata/fs"),
+		Index:              "index.html",
+		ContentTypeCharset: "UTF-8",
+	}))
+
+	resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/", nil))
+	require.NoError(t, err)
+	require.Equal(t, 200, resp.StatusCode)
+	require.Equal(t, "text/html; charset=UTF-8", resp.Header.Get("Content-Type"))
 }
