@@ -1,10 +1,9 @@
 package limiter
 
 import (
-	"fmt"
 	"time"
 
-	"github.com/gofiber/fiber/v2"
+	"github.com/gofiber/fiber/v3"
 )
 
 // Config defines the config for middleware.
@@ -12,19 +11,19 @@ type Config struct {
 	// Next defines a function to skip this middleware when returned true.
 	//
 	// Optional. Default: nil
-	Next func(c *fiber.Ctx) bool
+	Next func(c fiber.Ctx) bool
 
-	// Max number of recent connections during `Duration` seconds before sending a 429 response
+	// Max number of recent connections during `Expiration` seconds before sending a 429 response
 	//
 	// Default: 5
 	Max int
 
 	// KeyGenerator allows you to generate custom keys, by default c.IP() is used
 	//
-	// Default: func(c *fiber.Ctx) string {
+	// Default: func(c fiber.Ctx) string {
 	//   return c.IP()
 	// }
-	KeyGenerator func(*fiber.Ctx) string
+	KeyGenerator func(fiber.Ctx) string
 
 	// Expiration is the time on how long to keep records of requests in memory
 	//
@@ -33,36 +32,45 @@ type Config struct {
 
 	// LimitReached is called when a request hits the limit
 	//
-	// Default: func(c *fiber.Ctx) error {
+	// Default: func(c fiber.Ctx) error {
 	//   return c.SendStatus(fiber.StatusTooManyRequests)
 	// }
 	LimitReached fiber.Handler
+
+	// When set to true, requests with StatusCode >= 400 won't be counted.
+	//
+	// Default: false
+	SkipFailedRequests bool
+
+	// When set to true, requests with StatusCode < 400 won't be counted.
+	//
+	// Default: false
+	SkipSuccessfulRequests bool
 
 	// Store is used to store the state of the middleware
 	//
 	// Default: an in memory store for this process only
 	Storage fiber.Storage
 
-	// DEPRECATED: Use Expiration instead
-	Duration time.Duration
-
-	// DEPRECATED, use Storage instead
-	Store fiber.Storage
-
-	// DEPRECATED, use KeyGenerator instead
-	Key func(*fiber.Ctx) string
+	// LimiterMiddleware is the struct that implements a limiter middleware.
+	//
+	// Default: a new Fixed Window Rate Limiter
+	LimiterMiddleware LimiterHandler
 }
 
 // ConfigDefault is the default config
 var ConfigDefault = Config{
 	Max:        5,
 	Expiration: 1 * time.Minute,
-	KeyGenerator: func(c *fiber.Ctx) string {
+	KeyGenerator: func(c fiber.Ctx) string {
 		return c.IP()
 	},
-	LimitReached: func(c *fiber.Ctx) error {
+	LimitReached: func(c fiber.Ctx) error {
 		return c.SendStatus(fiber.StatusTooManyRequests)
 	},
+	SkipFailedRequests:     false,
+	SkipSuccessfulRequests: false,
+	LimiterMiddleware:      FixedWindow{},
 }
 
 // Helper function to set default values
@@ -76,18 +84,6 @@ func configDefault(config ...Config) Config {
 	cfg := config[0]
 
 	// Set default values
-	if int(cfg.Duration.Seconds()) > 0 {
-		fmt.Println("[LIMITER] Duration is deprecated, please use Expiration")
-		cfg.Expiration = cfg.Duration
-	}
-	if cfg.Key != nil {
-		fmt.Println("[LIMITER] Key is deprecated, please us KeyGenerator")
-		cfg.KeyGenerator = cfg.Key
-	}
-	if cfg.Store != nil {
-		fmt.Println("[LIMITER] Store is deprecated, please use Storage")
-		cfg.Storage = cfg.Store
-	}
 	if cfg.Next == nil {
 		cfg.Next = ConfigDefault.Next
 	}
@@ -102,6 +98,9 @@ func configDefault(config ...Config) Config {
 	}
 	if cfg.LimitReached == nil {
 		cfg.LimitReached = ConfigDefault.LimitReached
+	}
+	if cfg.LimiterMiddleware == nil {
+		cfg.LimiterMiddleware = ConfigDefault.LimiterMiddleware
 	}
 	return cfg
 }
