@@ -4,62 +4,91 @@ import (
 	"net/http/pprof"
 	"strings"
 
-	"github.com/gofiber/fiber/v2"
+	"github.com/gofiber/fiber/v3"
 	"github.com/valyala/fasthttp/fasthttpadaptor"
 )
 
-// Set pprof adaptors
-var (
-	pprofIndex        = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Index)
-	pprofCmdline      = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Cmdline)
-	pprofProfile      = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Profile)
-	pprofSymbol       = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Symbol)
-	pprofTrace        = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Trace)
-	pprofAllocs       = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Handler("allocs").ServeHTTP)
-	pprofBlock        = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Handler("block").ServeHTTP)
-	pprofGoroutine    = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Handler("goroutine").ServeHTTP)
-	pprofHeap         = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Handler("heap").ServeHTTP)
-	pprofMutex        = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Handler("mutex").ServeHTTP)
-	pprofThreadcreate = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Handler("threadcreate").ServeHTTP)
-)
-
 // New creates a new middleware handler
-func New() fiber.Handler {
+func New(config ...Config) fiber.Handler {
+	// Set default config
+	cfg := configDefault(config...)
+
+	// Set pprof adaptors
+	var (
+		pprofIndex        = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Index)
+		pprofCmdline      = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Cmdline)
+		pprofProfile      = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Profile)
+		pprofSymbol       = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Symbol)
+		pprofTrace        = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Trace)
+		pprofAllocs       = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Handler("allocs").ServeHTTP)
+		pprofBlock        = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Handler("block").ServeHTTP)
+		pprofGoroutine    = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Handler("goroutine").ServeHTTP)
+		pprofHeap         = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Handler("heap").ServeHTTP)
+		pprofMutex        = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Handler("mutex").ServeHTTP)
+		pprofThreadcreate = fasthttpadaptor.NewFastHTTPHandlerFunc(pprof.Handler("threadcreate").ServeHTTP)
+	)
+
+	// Construct actual prefix
+	prefix := cfg.Prefix + "/debug/pprof"
+
 	// Return new handler
-	return func(c *fiber.Ctx) error {
-		path := c.Path()
-		// We are only interested in /debug/pprof routes
-		if len(path) < 12 || !strings.HasPrefix(path, "/debug/pprof") {
+	return func(c fiber.Ctx) error {
+		// Don't execute middleware if Next returns true
+		if cfg.Next != nil && cfg.Next(c) {
 			return c.Next()
 		}
-		// Switch to original path without stripped slashes
+
+		path := c.Path()
+		// We are only interested in /debug/pprof routes
+		path, found := cutPrefix(path, prefix)
+		if !found {
+			return c.Next()
+		}
+		// Switch on trimmed path against constant strings
 		switch path {
-		case "/debug/pprof/":
+		case "/":
 			pprofIndex(c.Context())
-		case "/debug/pprof/cmdline":
+		case "/cmdline":
 			pprofCmdline(c.Context())
-		case "/debug/pprof/profile":
+		case "/profile":
 			pprofProfile(c.Context())
-		case "/debug/pprof/symbol":
+		case "/symbol":
 			pprofSymbol(c.Context())
-		case "/debug/pprof/trace":
+		case "/trace":
 			pprofTrace(c.Context())
-		case "/debug/pprof/allocs":
+		case "/allocs":
 			pprofAllocs(c.Context())
-		case "/debug/pprof/block":
+		case "/block":
 			pprofBlock(c.Context())
-		case "/debug/pprof/goroutine":
+		case "/goroutine":
 			pprofGoroutine(c.Context())
-		case "/debug/pprof/heap":
+		case "/heap":
 			pprofHeap(c.Context())
-		case "/debug/pprof/mutex":
+		case "/mutex":
 			pprofMutex(c.Context())
-		case "/debug/pprof/threadcreate":
+		case "/threadcreate":
 			pprofThreadcreate(c.Context())
 		default:
 			// pprof index only works with trailing slash
-			return c.Redirect("/debug/pprof/", 302)
+			if strings.HasSuffix(path, "/") {
+				path = strings.TrimRight(path, "/")
+			} else {
+				path = prefix + "/"
+			}
+
+			return c.Redirect().To(path)
 		}
 		return nil
 	}
+}
+
+// cutPrefix is a copy of [strings.CutPrefix] added in Go 1.20.
+// Remove this function when we drop support for Go 1.19.
+//
+//nolint:nonamedreturns // Align with its original form in std.
+func cutPrefix(s, prefix string) (after string, found bool) {
+	if !strings.HasPrefix(s, prefix) {
+		return s, false
+	}
+	return s[len(prefix):], true
 }
