@@ -1,7 +1,7 @@
-package recover
+package recover //nolint:predeclared // TODO: Rename to some non-builtin
 
 import (
-	"github.com/gofiber/fiber/v2"
+	"github.com/gofiber/fiber/v3"
 )
 
 // Config defines the config for middleware.
@@ -9,12 +9,24 @@ type Config struct {
 	// Next defines a function to skip this middleware when returned true.
 	//
 	// Optional. Default: nil
-	Next func(c *fiber.Ctx) bool
+	Next func(c fiber.Ctx) bool
+
+	// EnableStackTrace enables handling stack trace
+	//
+	// Optional. Default: false
+	EnableStackTrace bool
+
+	// StackTraceHandler defines a function to handle stack trace
+	//
+	// Optional. Default: defaultStackTraceHandler
+	StackTraceHandler func(c fiber.Ctx, e any)
 }
 
 // ConfigDefault is the default config
 var ConfigDefault = Config{
-	Next: nil,
+	Next:              nil,
+	EnableStackTrace:  false,
+	StackTraceHandler: defaultStackTraceHandler,
 }
 
 // Helper function to set default values
@@ -26,6 +38,10 @@ func configDefault(config ...Config) Config {
 
 	// Override default config
 	cfg := config[0]
+
+	if cfg.EnableStackTrace && cfg.StackTraceHandler == nil {
+		cfg.StackTraceHandler = defaultStackTraceHandler
+	}
 
 	return cfg
 }
