@@ -4,72 +4,73 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/gofiber/fiber/v2"
-	"github.com/gofiber/fiber/v2/utils"
+	"github.com/gofiber/fiber/v3"
+	"github.com/stretchr/testify/require"
 )
 
 // go test -run Test_RequestID
 func Test_RequestID(t *testing.T) {
+	t.Parallel()
 	app := fiber.New()
 
 	app.Use(New())
 
-	app.Get("/", func(c *fiber.Ctx) error {
+	app.Get("/", func(c fiber.Ctx) error {
 		return c.SendString("Hello, World 👋!")
 	})
 
-	resp, err := app.Test(httptest.NewRequest("GET", "/", nil))
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, fiber.StatusOK, resp.StatusCode)
+	resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/", nil))
+	require.NoError(t, err)
+	require.Equal(t, fiber.StatusOK, resp.StatusCode)
 
 	reqid := resp.Header.Get(fiber.HeaderXRequestID)
-	utils.AssertEqual(t, 36, len(reqid))
+	require.Len(t, reqid, 36)
 
-	req := httptest.NewRequest("GET", "/", nil)
+	req := httptest.NewRequest(fiber.MethodGet, "/", nil)
 	req.Header.Add(fiber.HeaderXRequestID, reqid)
 
 	resp, err = app.Test(req)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, fiber.StatusOK, resp.StatusCode)
-	utils.AssertEqual(t, reqid, resp.Header.Get(fiber.HeaderXRequestID))
+	require.NoError(t, err)
+	require.Equal(t, fiber.StatusOK, resp.StatusCode)
+	require.Equal(t, reqid, resp.Header.Get(fiber.HeaderXRequestID))
 }
 
 // go test -run Test_RequestID_Next
 func Test_RequestID_Next(t *testing.T) {
+	t.Parallel()
 	app := fiber.New()
 	app.Use(New(Config{
-		Next: func(_ *fiber.Ctx) bool {
+		Next: func(_ fiber.Ctx) bool {
 			return true
 		},
 	}))
 
-	resp, err := app.Test(httptest.NewRequest("GET", "/", nil))
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, resp.Header.Get(fiber.HeaderXRequestID), "")
-	utils.AssertEqual(t, fiber.StatusNotFound, resp.StatusCode)
+	resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/", nil))
+	require.NoError(t, err)
+	require.Empty(t, resp.Header.Get(fiber.HeaderXRequestID))
+	require.Equal(t, fiber.StatusNotFound, resp.StatusCode)
 }
 
 // go test -run Test_RequestID_Locals
-func Test_RequestID_Locals(t *testing.T) {
-	reqId := "ThisIsARequestId"
-	ctxKey := "ThisIsAContextKey"
+func Test_RequestID_FromContext(t *testing.T) {
+	t.Parallel()
+	reqID := "ThisIsARequestId"
 
 	app := fiber.New()
 	app.Use(New(Config{
 		Generator: func() string {
-			return reqId
+			return reqID
 		},
-		ContextKey: ctxKey,
 	}))
 
 	var ctxVal string
 
-	app.Use(func(c *fiber.Ctx) error {
-		ctxVal = c.Locals(ctxKey).(string)
+	app.Use(func(c fiber.Ctx) error {
+		ctxVal = FromContext(c)
 		return c.Next()
 	})
 
-	_, err := app.Test(httptest.NewRequest("GET", "/", nil))
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, reqId, ctxVal)
+	_, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/", nil))
+	require.NoError(t, err)
+	require.Equal(t, reqID, ctxVal)
 }
