@@ -1,10 +1,11 @@
 package session
 
 import (
+	"strings"
 	"time"
 
-	"github.com/gofiber/fiber/v2"
-	"github.com/gofiber/fiber/v2/utils"
+	"github.com/gofiber/fiber/v3"
+	"github.com/gofiber/utils/v2"
 )
 
 // Config defines the config for middleware.
@@ -17,40 +18,63 @@ type Config struct {
 	// Optional. Default value memory.New()
 	Storage fiber.Storage
 
-	// Name of the session cookie. This cookie will store session key.
-	// Optional. Default value "session_id".
-	CookieName string
+	// KeyLookup is a string in the form of "<source>:<name>" that is used
+	// to extract session id from the request.
+	// Possible values: "header:<name>", "query:<name>" or "cookie:<name>"
+	// Optional. Default value "cookie:session_id".
+	KeyLookup string
 
-	// Domain of the CSRF cookie.
+	// Domain of the cookie.
 	// Optional. Default value "".
 	CookieDomain string
 
-	// Path of the CSRF cookie.
+	// Path of the cookie.
 	// Optional. Default value "".
 	CookiePath string
 
-	// Indicates if CSRF cookie is secure.
+	// Indicates if cookie is secure.
 	// Optional. Default value false.
 	CookieSecure bool
 
-	// Indicates if CSRF cookie is HTTP only.
+	// Indicates if cookie is HTTP only.
 	// Optional. Default value false.
 	CookieHTTPOnly bool
 
-	// Indicates if CSRF cookie is HTTP only.
-	// Optional. Default value false.
+	// Value of SameSite cookie.
+	// Optional. Default value "Lax".
 	CookieSameSite string
 
+	// Decides whether cookie should last for only the browser sesison.
+	// Ignores Expiration if set to true
+	// Optional. Default value false.
+	CookieSessionOnly bool
+
 	// KeyGenerator generates the session key.
-	// Optional. Default value utils.UUID
+	// Optional. Default value utils.UUIDv4
 	KeyGenerator func() string
+
+	// Source defines where to obtain the session id
+	source Source
+
+	// The session name
+	sessionName string
 }
+
+type Source string
+
+const (
+	SourceCookie   Source = "cookie"
+	SourceHeader   Source = "header"
+	SourceURLQuery Source = "query"
+)
 
 // ConfigDefault is the default config
 var ConfigDefault = Config{
 	Expiration:   24 * time.Hour,
-	CookieName:   "session_id",
-	KeyGenerator: utils.UUID,
+	KeyLookup:    "cookie:session_id",
+	KeyGenerator: utils.UUIDv4,
+	source:       "cookie",
+	sessionName:  "session_id",
 }
 
 // Helper function to set default values
@@ -67,8 +91,29 @@ func configDefault(config ...Config) Config {
 	if int(cfg.Expiration.Seconds()) <= 0 {
 		cfg.Expiration = ConfigDefault.Expiration
 	}
+	if cfg.KeyLookup == "" {
+		cfg.KeyLookup = ConfigDefault.KeyLookup
+	}
 	if cfg.KeyGenerator == nil {
 		cfg.KeyGenerator = ConfigDefault.KeyGenerator
 	}
+
+	selectors := strings.Split(cfg.KeyLookup, ":")
+	const numSelectors = 2
+	if len(selectors) != numSelectors {
+		panic("[session] KeyLookup must in the form of <source>:<name>")
+	}
+	switch Source(selectors[0]) {
+	case SourceCookie:
+		cfg.source = SourceCookie
+	case SourceHeader:
+		cfg.source = SourceHeader
+	case SourceURLQuery:
+		cfg.source = SourceURLQuery
+	default:
+		panic("[session] source is not supported")
+	}
+	cfg.sessionName = selectors[1]
+
 	return cfg
 }
