@@ -1,19 +1,25 @@
 package timeout
 
 import (
-	"fmt"
+	"context"
+	"errors"
 	"sync"
 	"time"
 
-	"github.com/gofiber/fiber/v2"
+	"github.com/gofiber/fiber/v3/log"
+
+	"github.com/gofiber/fiber/v3"
 )
 
 var once sync.Once
 
-// New wraps a handler and aborts the process of the handler if the timeout is reached
+// New wraps a handler and aborts the process of the handler if the timeout is reached.
+//
+// Deprecated: This implementation contains data race issues. Use NewWithContext instead.
+// Find documentation and sample usage on https://docs.gofiber.io/api/middleware/timeout
 func New(handler fiber.Handler, timeout time.Duration) fiber.Handler {
 	once.Do(func() {
-		fmt.Println("[Warning] timeout contains data race issues, not ready for production!")
+		log.Warn("[TIMEOUT] timeout contains data race issues, not ready for production!")
 	})
 
 	if timeout <= 0 {
@@ -21,14 +27,18 @@ func New(handler fiber.Handler, timeout time.Duration) fiber.Handler {
 	}
 
 	// logic is from fasthttp.TimeoutWithCodeHandler https://github.com/valyala/fasthttp/blob/master/server.go#L418
-	return func(ctx *fiber.Ctx) error {
+	return func(c fiber.Ctx) error {
 		ch := make(chan struct{}, 1)
 
 		go func() {
 			defer func() {
-				_ = recover()
+				if err := recover(); err != nil {
+					log.Errorf("[TIMEOUT] recover error %v", err)
+				}
 			}()
-			_ = handler(ctx)
+			if err := handler(c); err != nil {
+				log.Errorf("[TIMEOUT] handler error %v", err)
+			}
 			ch <- struct{}{}
 		}()
 
@@ -38,6 +48,27 @@ func New(handler fiber.Handler, timeout time.Duration) fiber.Handler {
 			return fiber.ErrRequestTimeout
 		}
 
+		return nil
+	}
+}
+
+// NewWithContext implementation of timeout middleware. Set custom errors(context.DeadlineExceeded vs) for get fiber.ErrRequestTimeout response.
+func NewWithContext(h fiber.Handler, t time.Duration, tErrs ...error) fiber.Handler {
+	return func(ctx fiber.Ctx) error {
+		timeoutContext, cancel := context.WithTimeout(ctx.UserContext(), t)
+		defer cancel()
+		ctx.SetUserContext(timeoutContext)
+		if err := h(ctx); err != nil {
+			if errors.Is(err, context.DeadlineExceeded) {
+				return fiber.ErrRequestTimeout
+			}
+			for i := range tErrs {
+				if errors.Is(err, tErrs[i]) {
+					return fiber.ErrRequestTimeout
+				}
+			}
+			return err
+		}
 		return nil
 	}
 }
