@@ -1,55 +1,88 @@
 package timeout
 
-// // go test -run Test_Middleware_Timeout
-// func Test_Middleware_Timeout(t *testing.T) {
-// 	app := fiber.New(fiber.Config{DisableStartupMessage: true})
+import (
+	"context"
+	"errors"
+	"fmt"
+	"net/http/httptest"
+	"testing"
+	"time"
 
-// 	h := New(func(c *fiber.Ctx) error {
-// 		sleepTime, _ := time.ParseDuration(c.Params("sleepTime") + "ms")
-// 		time.Sleep(sleepTime)
-// 		return c.SendString("After " + c.Params("sleepTime") + "ms sleeping")
-// 	}, 5*time.Millisecond)
-// 	app.Get("/test/:sleepTime", h)
+	"github.com/gofiber/fiber/v3"
+	"github.com/stretchr/testify/require"
+)
 
-// 	testTimeout := func(timeoutStr string) {
-// 		resp, err := app.Test(httptest.NewRequest("GET", "/test/"+timeoutStr, nil))
-// 		utils.AssertEqual(t, nil, err, "app.Test(req)")
-// 		utils.AssertEqual(t, fiber.StatusRequestTimeout, resp.StatusCode, "Status code")
+// go test -run Test_WithContextTimeout
+func Test_WithContextTimeout(t *testing.T) {
+	t.Parallel()
+	// fiber instance
+	app := fiber.New()
+	h := NewWithContext(func(c fiber.Ctx) error {
+		sleepTime, err := time.ParseDuration(c.Params("sleepTime") + "ms")
+		require.NoError(t, err)
+		if err := sleepWithContext(c.UserContext(), sleepTime, context.DeadlineExceeded); err != nil {
+			return fmt.Errorf("%w: l2 wrap", fmt.Errorf("%w: l1 wrap ", err))
+		}
+		return nil
+	}, 100*time.Millisecond)
+	app.Get("/test/:sleepTime", h)
+	testTimeout := func(timeoutStr string) {
+		resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/test/"+timeoutStr, nil))
+		require.NoError(t, err, "app.Test(req)")
+		require.Equal(t, fiber.StatusRequestTimeout, resp.StatusCode, "Status code")
+	}
+	testSucces := func(timeoutStr string) {
+		resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/test/"+timeoutStr, nil))
+		require.NoError(t, err, "app.Test(req)")
+		require.Equal(t, fiber.StatusOK, resp.StatusCode, "Status code")
+	}
+	testTimeout("300")
+	testTimeout("500")
+	testSucces("50")
+	testSucces("30")
+}
 
-// 		body, err := ioutil.ReadAll(resp.Body)
-// 		utils.AssertEqual(t, nil, err)
-// 		utils.AssertEqual(t, "Request Timeout", string(body))
-// 	}
-// 	testSucces := func(timeoutStr string) {
-// 		resp, err := app.Test(httptest.NewRequest("GET", "/test/"+timeoutStr, nil))
-// 		utils.AssertEqual(t, nil, err, "app.Test(req)")
-// 		utils.AssertEqual(t, fiber.StatusOK, resp.StatusCode, "Status code")
+var ErrFooTimeOut = errors.New("foo context canceled")
 
-// 		body, err := ioutil.ReadAll(resp.Body)
-// 		utils.AssertEqual(t, nil, err)
-// 		utils.AssertEqual(t, "After "+timeoutStr+"ms sleeping", string(body))
-// 	}
+// go test -run Test_WithContextTimeoutWithCustomError
+func Test_WithContextTimeoutWithCustomError(t *testing.T) {
+	t.Parallel()
+	// fiber instance
+	app := fiber.New()
+	h := NewWithContext(func(c fiber.Ctx) error {
+		sleepTime, err := time.ParseDuration(c.Params("sleepTime") + "ms")
+		require.NoError(t, err)
+		if err := sleepWithContext(c.UserContext(), sleepTime, ErrFooTimeOut); err != nil {
+			return fmt.Errorf("%w: execution error", err)
+		}
+		return nil
+	}, 100*time.Millisecond, ErrFooTimeOut)
+	app.Get("/test/:sleepTime", h)
+	testTimeout := func(timeoutStr string) {
+		resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/test/"+timeoutStr, nil))
+		require.NoError(t, err, "app.Test(req)")
+		require.Equal(t, fiber.StatusRequestTimeout, resp.StatusCode, "Status code")
+	}
+	testSucces := func(timeoutStr string) {
+		resp, err := app.Test(httptest.NewRequest(fiber.MethodGet, "/test/"+timeoutStr, nil))
+		require.NoError(t, err, "app.Test(req)")
+		require.Equal(t, fiber.StatusOK, resp.StatusCode, "Status code")
+	}
+	testTimeout("300")
+	testTimeout("500")
+	testSucces("50")
+	testSucces("30")
+}
 
-// 	testTimeout("15")
-// 	testSucces("2")
-// 	testTimeout("30")
-// 	testSucces("3")
-// }
-
-// // go test -run -v Test_Timeout_Panic
-// func Test_Timeout_Panic(t *testing.T) {
-// 	app := fiber.New(fiber.Config{DisableStartupMessage: true})
-
-// 	app.Get("/panic", recover.New(), New(func(c *fiber.Ctx) error {
-// 		c.Set("dummy", "this should not be here")
-// 		panic("panic in timeout handler")
-// 	}, 5*time.Millisecond))
-
-// 	resp, err := app.Test(httptest.NewRequest("GET", "/panic", nil))
-// 	utils.AssertEqual(t, nil, err, "app.Test(req)")
-// 	utils.AssertEqual(t, fiber.StatusRequestTimeout, resp.StatusCode, "Status code")
-
-// 	body, err := ioutil.ReadAll(resp.Body)
-// 	utils.AssertEqual(t, nil, err)
-// 	utils.AssertEqual(t, "Request Timeout", string(body))
-// }
+func sleepWithContext(ctx context.Context, d time.Duration, te error) error {
+	timer := time.NewTimer(d)
+	select {
+	case <-ctx.Done():
+		if !timer.Stop() {
+			<-timer.C
+		}
+		return te
+	case <-timer.C:
+	}
+	return nil
+}
