@@ -8,15 +8,16 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/gofiber/fiber/v2/utils"
+	"github.com/stretchr/testify/require"
 )
 
 // go test -race -run Test_Path_parseRoute
 func Test_Path_parseRoute(t *testing.T) {
+	t.Parallel()
 	var rp routeParser
 
 	rp = parseRoute("/shop/product/::filter/color::color/size::size")
-	utils.AssertEqual(t, routeParser{
+	require.Equal(t, routeParser{
 		segs: []*routeSegment{
 			{Const: "/shop/product/:", Length: 15},
 			{IsParam: true, ParamName: "filter", ComparePart: "/color:", PartCount: 1},
@@ -29,7 +30,7 @@ func Test_Path_parseRoute(t *testing.T) {
 	}, rp)
 
 	rp = parseRoute("/api/v1/:param/abc/*")
-	utils.AssertEqual(t, routeParser{
+	require.Equal(t, routeParser{
 		segs: []*routeSegment{
 			{Const: "/api/v1/", Length: 8},
 			{IsParam: true, ParamName: "param", ComparePart: "/abc", PartCount: 1},
@@ -40,8 +41,39 @@ func Test_Path_parseRoute(t *testing.T) {
 		wildCardCount: 1,
 	}, rp)
 
+	rp = parseRoute("/v1/some/resource/name\\:customVerb")
+	require.Equal(t, routeParser{
+		segs: []*routeSegment{
+			{Const: "/v1/some/resource/name:customVerb", Length: 33, IsLast: true},
+		},
+		params: nil,
+	}, rp)
+
+	rp = parseRoute("/v1/some/resource/:name\\:customVerb")
+	require.Equal(t, routeParser{
+		segs: []*routeSegment{
+			{Const: "/v1/some/resource/", Length: 18},
+			{IsParam: true, ParamName: "name", ComparePart: ":customVerb", PartCount: 1},
+			{Const: ":customVerb", Length: 11, IsLast: true},
+		},
+		params: []string{"name"},
+	}, rp)
+
+	// heavy test with escaped charaters
+	rp = parseRoute("/v1/some/resource/name\\\\:customVerb?\\?/:param/*")
+	require.Equal(t, routeParser{
+		segs: []*routeSegment{
+			{Const: "/v1/some/resource/name:customVerb??/", Length: 36},
+			{IsParam: true, ParamName: "param", ComparePart: "/", PartCount: 1},
+			{Const: "/", Length: 1, HasOptionalSlash: true},
+			{IsParam: true, ParamName: "*1", IsGreedy: true, IsOptional: true, IsLast: true},
+		},
+		params:        []string{"param", "*1"},
+		wildCardCount: 1,
+	}, rp)
+
 	rp = parseRoute("/api/*/:param/:param2")
-	utils.AssertEqual(t, routeParser{
+	require.Equal(t, routeParser{
 		segs: []*routeSegment{
 			{Const: "/api/", Length: 5, HasOptionalSlash: true},
 			{IsParam: true, ParamName: "*1", IsGreedy: true, IsOptional: true, ComparePart: "/", PartCount: 2},
@@ -55,7 +87,7 @@ func Test_Path_parseRoute(t *testing.T) {
 	}, rp)
 
 	rp = parseRoute("/test:optional?:optional2?")
-	utils.AssertEqual(t, routeParser{
+	require.Equal(t, routeParser{
 		segs: []*routeSegment{
 			{Const: "/test", Length: 5},
 			{IsParam: true, ParamName: "optional", IsOptional: true, Length: 1},
@@ -65,7 +97,7 @@ func Test_Path_parseRoute(t *testing.T) {
 	}, rp)
 
 	rp = parseRoute("/config/+.json")
-	utils.AssertEqual(t, routeParser{
+	require.Equal(t, routeParser{
 		segs: []*routeSegment{
 			{Const: "/config/", Length: 8},
 			{IsParam: true, ParamName: "+1", IsGreedy: true, IsOptional: false, ComparePart: ".json", PartCount: 1},
@@ -76,7 +108,7 @@ func Test_Path_parseRoute(t *testing.T) {
 	}, rp)
 
 	rp = parseRoute("/api/:day.:month?.:year?")
-	utils.AssertEqual(t, routeParser{
+	require.Equal(t, routeParser{
 		segs: []*routeSegment{
 			{Const: "/api/", Length: 5},
 			{IsParam: true, ParamName: "day", IsOptional: false, ComparePart: ".", PartCount: 2},
@@ -89,7 +121,7 @@ func Test_Path_parseRoute(t *testing.T) {
 	}, rp)
 
 	rp = parseRoute("/*v1*/proxy")
-	utils.AssertEqual(t, routeParser{
+	require.Equal(t, routeParser{
 		segs: []*routeSegment{
 			{Const: "/", Length: 1, HasOptionalSlash: true},
 			{IsParam: true, ParamName: "*1", IsGreedy: true, IsOptional: true, ComparePart: "v1", PartCount: 1},
@@ -105,363 +137,132 @@ func Test_Path_parseRoute(t *testing.T) {
 // go test -race -run Test_Path_matchParams
 func Test_Path_matchParams(t *testing.T) {
 	t.Parallel()
-	type testparams struct {
-		url          string
-		params       []string
-		match        bool
-		partialCheck bool
-	}
 	var ctxParams [maxParams]string
-	testCase := func(r string, cases []testparams) {
-		parser := parseRoute(r)
-		for _, c := range cases {
+	testCaseFn := func(testCollection routeCaseCollection) {
+		parser := parseRoute(testCollection.pattern)
+		for _, c := range testCollection.testCases {
 			match := parser.getMatch(c.url, c.url, &ctxParams, c.partialCheck)
-			utils.AssertEqual(t, c.match, match, fmt.Sprintf("route: '%s', url: '%s'", r, c.url))
+			require.Equal(t, c.match, match, fmt.Sprintf("route: '%s', url: '%s'", testCollection.pattern, c.url))
 			if match && len(c.params) > 0 {
-				utils.AssertEqual(t, c.params[0:len(c.params)-1], ctxParams[0:len(c.params)-1], fmt.Sprintf("route: '%s', url: '%s'", r, c.url))
+				require.Equal(t, c.params[0:len(c.params)], ctxParams[0:len(c.params)], fmt.Sprintf("route: '%s', url: '%s'", testCollection.pattern, c.url))
 			}
 		}
 	}
-	testCase("/api/v1/:param/*", []testparams{
-		{url: "/api/v1/entity", params: []string{"entity", ""}, match: true},
-		{url: "/api/v1/entity/", params: []string{"entity", ""}, match: true},
-		{url: "/api/v1/entity/1", params: []string{"entity", "1"}, match: true},
-		{url: "/api/v", params: nil, match: false},
-		{url: "/api/v2", params: nil, match: false},
-		{url: "/api/v1/", params: nil, match: false},
-	})
-	testCase("/api/v1/:param/+", []testparams{
-		{url: "/api/v1/entity", params: nil, match: false},
-		{url: "/api/v1/entity/", params: nil, match: false},
-		{url: "/api/v1/entity/1", params: []string{"entity", "1"}, match: true},
-		{url: "/api/v", params: nil, match: false},
-		{url: "/api/v2", params: nil, match: false},
-		{url: "/api/v1/", params: nil, match: false},
-	})
-	testCase("/api/v1/:param?", []testparams{
-		{url: "/api/v1", params: []string{""}, match: true},
-		{url: "/api/v1/", params: []string{""}, match: true},
-		{url: "/api/v1/optional", params: []string{"optional"}, match: true},
-		{url: "/api/v", params: nil, match: false},
-		{url: "/api/v2", params: nil, match: false},
-		{url: "/api/xyz", params: nil, match: false},
-	})
-	testCase("/api/v1/*", []testparams{
-		{url: "/api/v1", params: []string{""}, match: true},
-		{url: "/api/v1/", params: []string{""}, match: true},
-		{url: "/api/v1/entity", params: []string{"entity"}, match: true},
-		{url: "/api/v1/entity/1/2", params: []string{"entity/1/2"}, match: true},
-		{url: "/api/v1/Entity/1/2", params: []string{"Entity/1/2"}, match: true},
-		{url: "/api/v", params: nil, match: false},
-		{url: "/api/v2", params: nil, match: false},
-		{url: "/api/abc", params: nil, match: false},
-	})
-	testCase("/api/v1/:param", []testparams{
-		{url: "/api/v1/entity", params: []string{"entity"}, match: true},
-		{url: "/api/v1/entity/8728382", params: nil, match: false},
-		{url: "/api/v1", params: nil, match: false},
-		{url: "/api/v1/", params: nil, match: false},
-	})
-	testCase("/api/v1/:param-:param2", []testparams{
-		{url: "/api/v1/entity-entity2", params: []string{"entity", "entity2"}, match: true},
-		{url: "/api/v1/entity/8728382", params: nil, match: false},
-		{url: "/api/v1/entity-8728382", params: []string{"entity", "8728382"}, match: true},
-		{url: "/api/v1", params: nil, match: false},
-		{url: "/api/v1/", params: nil, match: false},
-	})
-	testCase("/api/v1/:filename.:extension", []testparams{
-		{url: "/api/v1/test.pdf", params: []string{"test", "pdf"}, match: true},
-		{url: "/api/v1/test/pdf", params: nil, match: false},
-		{url: "/api/v1/test-pdf", params: nil, match: false},
-		{url: "/api/v1/test_pdf", params: nil, match: false},
-		{url: "/api/v1", params: nil, match: false},
-		{url: "/api/v1/", params: nil, match: false},
-	})
-	testCase("/api/v1/const", []testparams{
-		{url: "/api/v1/const", params: []string{}, match: true},
-		{url: "/api/v1", params: nil, match: false},
-		{url: "/api/v1/", params: nil, match: false},
-		{url: "/api/v1/something", params: nil, match: false},
-	})
-	testCase("/api/:param/fixedEnd", []testparams{
-		{url: "/api/abc/fixedEnd", params: []string{"abc"}, match: true},
-		{url: "/api/abc/def/fixedEnd", params: nil, match: false},
-	})
-	testCase("/shop/product/::filter/color::color/size::size", []testparams{
-		{url: "/shop/product/:test/color:blue/size:xs", params: []string{"test", "blue", "xs"}, match: true},
-		{url: "/shop/product/test/color:blue/size:xs", params: nil, match: false},
-	})
-	testCase("/::param?", []testparams{
-		{url: "/:hello", params: []string{"hello"}, match: true},
-		{url: "/:", params: []string{""}, match: true},
-		{url: "/", params: nil, match: false},
-	})
-	// successive parameters, each take one character and the last parameter gets everything
-	testCase("/test:sign:param", []testparams{
-		{url: "/test-abc", params: []string{"-", "abc"}, match: true},
-		{url: "/test", params: nil, match: false},
-	})
-	// optional parameters are not greedy
-	testCase("/:param1:param2?:param3", []testparams{
-		{url: "/abbbc", params: []string{"a", "b", "bbc"}, match: true},
-		//{url: "/ac", params: []string{"a", "", "c"}, match: true}, // TODO: fix it
-		{url: "/test", params: []string{"t", "e", "st"}, match: true},
-	})
-	testCase("/test:optional?:mandatory", []testparams{
-		//{url: "/testo", params: []string{"", "o"}, match: true}, // TODO: fix it
-		{url: "/testoaaa", params: []string{"o", "aaa"}, match: true},
-		{url: "/test", params: nil, match: false},
-	})
-	testCase("/test:optional?:optional2?", []testparams{
-		{url: "/testo", params: []string{"o", ""}, match: true},
-		{url: "/testoaaa", params: []string{"o", "aaa"}, match: true},
-		{url: "/test", params: []string{"", ""}, match: true},
-		{url: "/tes", params: nil, match: false},
-	})
-	testCase("/foo:param?bar", []testparams{
-		{url: "/foofaselbar", params: []string{"fasel"}, match: true},
-		{url: "/foobar", params: []string{""}, match: true},
-		{url: "/fooba", params: nil, match: false},
-		{url: "/fobar", params: nil, match: false},
-	})
-	testCase("/foo*bar", []testparams{
-		{url: "/foofaselbar", params: []string{"fasel"}, match: true},
-		{url: "/foobar", params: []string{""}, match: true},
-		{url: "/", params: nil, match: false},
-	})
-	testCase("/foo+bar", []testparams{
-		{url: "/foofaselbar", params: []string{"fasel"}, match: true},
-		{url: "/foobar", params: nil, match: false},
-		{url: "/", params: nil, match: false},
-	})
-	testCase("/a*cde*g/", []testparams{
-		{url: "/abbbcdefffg", params: []string{"bbb", "fff"}, match: true},
-		{url: "/acdeg", params: []string{"", ""}, match: true},
-		{url: "/", params: nil, match: false},
-	})
-	testCase("/*v1*/proxy", []testparams{
-		{url: "/customer/v1/cart/proxy", params: []string{"customer/", "/cart"}, match: true},
-		{url: "/v1/proxy", params: []string{"", ""}, match: true},
-		{url: "/v1/", params: nil, match: false},
-	})
-	// successive wildcard -> first wildcard is greedy
-	testCase("/foo***bar", []testparams{
-		{url: "/foo*abar", params: []string{"*a", "", ""}, match: true},
-		{url: "/foo*bar", params: []string{"*", "", ""}, match: true},
-		{url: "/foobar", params: []string{"", "", ""}, match: true},
-		{url: "/fooba", params: nil, match: false},
-	})
-	// chars in front of an parameter
-	testCase("/name::name", []testparams{
-		{url: "/name:john", params: []string{"john"}, match: true},
-	})
-	testCase("/@:name", []testparams{
-		{url: "/@john", params: []string{"john"}, match: true},
-	})
-	testCase("/-:name", []testparams{
-		{url: "/-john", params: []string{"john"}, match: true},
-	})
-	testCase("/.:name", []testparams{
-		{url: "/.john", params: []string{"john"}, match: true},
-	})
-	testCase("/api/v1/:param/abc/*", []testparams{
-		{url: "/api/v1/well/abc/wildcard", params: []string{"well", "wildcard"}, match: true},
-		{url: "/api/v1/well/abc/", params: []string{"well", ""}, match: true},
-		{url: "/api/v1/well/abc", params: []string{"well", ""}, match: true},
-		{url: "/api/v1/well/ttt", params: nil, match: false},
-	})
-	testCase("/api/:day/:month?/:year?", []testparams{
-		{url: "/api/1", params: []string{"1", "", ""}, match: true},
-		{url: "/api/1/", params: []string{"1", "", ""}, match: true},
-		{url: "/api/1//", params: []string{"1", "", ""}, match: true},
-		{url: "/api/1/-/", params: []string{"1", "-", ""}, match: true},
-		{url: "/api/1-", params: []string{"1-", "", ""}, match: true},
-		{url: "/api/1.", params: []string{"1.", "", ""}, match: true},
-		{url: "/api/1/2", params: []string{"1", "2", ""}, match: true},
-		{url: "/api/1/2/3", params: []string{"1", "2", "3"}, match: true},
-		{url: "/api/", params: nil, match: false},
-	})
-	testCase("/api/:day.:month?.:year?", []testparams{
-		{url: "/api/1", params: nil, match: false},
-		{url: "/api/1/", params: nil, match: false},
-		{url: "/api/1.", params: nil, match: false},
-		{url: "/api/1..", params: []string{"1", "", ""}, match: true},
-		{url: "/api/1.2", params: nil, match: false},
-		{url: "/api/1.2.", params: []string{"1", "2", ""}, match: true},
-		{url: "/api/1.2.3", params: []string{"1", "2", "3"}, match: true},
-		{url: "/api/", params: nil, match: false},
-	})
-	testCase("/api/:day-:month?-:year?", []testparams{
-		{url: "/api/1", params: nil, match: false},
-		{url: "/api/1/", params: nil, match: false},
-		{url: "/api/1-", params: nil, match: false},
-		{url: "/api/1--", params: []string{"1", "", ""}, match: true},
-		{url: "/api/1-/", params: nil, match: false},
-		//{url: "/api/1-/-", params: nil, match: false}, // TODO: fix this part
-		{url: "/api/1-2", params: nil, match: false},
-		{url: "/api/1-2-", params: []string{"1", "2", ""}, match: true},
-		{url: "/api/1-2-3", params: []string{"1", "2", "3"}, match: true},
-		{url: "/api/", params: nil, match: false},
-	})
-	testCase("/api/*", []testparams{
-		{url: "/api/", params: []string{""}, match: true},
-		{url: "/api/joker", params: []string{"joker"}, match: true},
-		{url: "/api", params: []string{""}, match: true},
-		{url: "/api/v1/entity", params: []string{"v1/entity"}, match: true},
-		{url: "/api2/v1/entity", params: nil, match: false},
-		{url: "/api_ignore/v1/entity", params: nil, match: false},
-	})
-	testCase("/partialCheck/foo/bar/:param", []testparams{
-		{url: "/partialCheck/foo/bar/test", params: []string{"test"}, match: true, partialCheck: true},
-		{url: "/partialCheck/foo/bar/test/test2", params: []string{"test"}, match: true, partialCheck: true},
-		{url: "/partialCheck/foo/bar", params: nil, match: false, partialCheck: true},
-		{url: "/partiaFoo", params: nil, match: false, partialCheck: true},
-	})
-	testCase("/", []testparams{
-		{url: "/api", params: nil, match: false},
-		{url: "", params: []string{}, match: true},
-		{url: "/", params: []string{}, match: true},
-	})
-	testCase("/config/abc.json", []testparams{
-		{url: "/config/abc.json", params: []string{}, match: true},
-		{url: "config/abc.json", params: nil, match: false},
-		{url: "/config/efg.json", params: nil, match: false},
-		{url: "/config", params: nil, match: false},
-	})
-	testCase("/config/*.json", []testparams{
-		{url: "/config/abc.json", params: []string{"abc"}, match: true},
-		{url: "/config/efg.json", params: []string{"efg"}, match: true},
-		{url: "/config/.json", params: []string{""}, match: true},
-		{url: "/config/efg.csv", params: nil, match: false},
-		{url: "config/abc.json", params: nil, match: false},
-		{url: "/config", params: nil, match: false},
-	})
-	testCase("/config/+.json", []testparams{
-		{url: "/config/abc.json", params: []string{"abc"}, match: true},
-		{url: "/config/.json", params: nil, match: false},
-		{url: "/config/efg.json", params: []string{"efg"}, match: true},
-		{url: "/config/efg.csv", params: nil, match: false},
-		{url: "config/abc.json", params: nil, match: false},
-		{url: "/config", params: nil, match: false},
-	})
-	testCase("/xyz", []testparams{
-		{url: "xyz", params: nil, match: false},
-		{url: "xyz/", params: nil, match: false},
-	})
-	testCase("/api/*/:param?", []testparams{
-		{url: "/api/", params: []string{"", ""}, match: true},
-		{url: "/api/joker", params: []string{"joker", ""}, match: true},
-		{url: "/api/joker/batman", params: []string{"joker", "batman"}, match: true},
-		{url: "/api/joker//batman", params: []string{"joker/", "batman"}, match: true},
-		{url: "/api/joker/batman/robin", params: []string{"joker/batman", "robin"}, match: true},
-		{url: "/api/joker/batman/robin/1", params: []string{"joker/batman/robin", "1"}, match: true},
-		{url: "/api/joker/batman/robin/1/", params: []string{"joker/batman/robin/1", ""}, match: true},
-		{url: "/api/joker-batman/robin/1", params: []string{"joker-batman/robin", "1"}, match: true},
-		{url: "/api/joker-batman-robin/1", params: []string{"joker-batman-robin", "1"}, match: true},
-		{url: "/api/joker-batman-robin-1", params: []string{"joker-batman-robin-1", ""}, match: true},
-		{url: "/api", params: []string{"", ""}, match: true},
-	})
-	testCase("/api/*/:param", []testparams{
-		{url: "/api/test/abc", params: []string{"test", "abc"}, match: true},
-		{url: "/api/joker/batman", params: []string{"joker", "batman"}, match: true},
-		{url: "/api/joker/batman/robin", params: []string{"joker/batman", "robin"}, match: true},
-		{url: "/api/joker/batman/robin/1", params: []string{"joker/batman/robin", "1"}, match: true},
-		{url: "/api/joker/batman-robin/1", params: []string{"joker/batman-robin", "1"}, match: true},
-		{url: "/api/joker-batman-robin-1", params: nil, match: false},
-		{url: "/api", params: nil, match: false},
-	})
-	testCase("/api/+/:param", []testparams{
-		{url: "/api/test/abc", params: []string{"test", "abc"}, match: true},
-		{url: "/api/joker/batman/robin/1", params: []string{"joker/batman/robin", "1"}, match: true},
-		{url: "/api/joker", params: nil, match: false},
-		{url: "/api", params: nil, match: false},
-	})
-	testCase("/api/*/:param/:param2", []testparams{
-		{url: "/api/test/abc/1", params: []string{"test", "abc", "1"}, match: true},
-		{url: "/api/joker/batman", params: nil, match: false},
-		{url: "/api/joker/batman-robin/1", params: []string{"joker", "batman-robin", "1"}, match: true},
-		{url: "/api/joker-batman-robin-1", params: nil, match: false},
-		{url: "/api/test/abc", params: nil, match: false},
-		{url: "/api/joker/batman/robin", params: []string{"joker", "batman", "robin"}, match: true},
-		{url: "/api/joker/batman/robin/1", params: []string{"joker/batman", "robin", "1"}, match: true},
-		{url: "/api/joker/batman/robin/1/2", params: []string{"joker/batman/robin", "1", "2"}, match: true},
-		{url: "/api", params: nil, match: false},
-		{url: "/api/:test", params: nil, match: false},
-	})
+	for _, testCaseCollection := range routeTestCases {
+		testCaseFn(testCaseCollection)
+	}
+}
+
+// go test -race -run Test_RoutePatternMatch
+func Test_RoutePatternMatch(t *testing.T) {
+	t.Parallel()
+	testCaseFn := func(pattern string, cases []routeTestCase) {
+		for _, c := range cases {
+			// skip all cases for partial checks
+			if c.partialCheck {
+				continue
+			}
+			match := RoutePatternMatch(c.url, pattern)
+			require.Equal(t, c.match, match, fmt.Sprintf("route: '%s', url: '%s'", pattern, c.url))
+		}
+	}
+	for _, testCase := range routeTestCases {
+		testCaseFn(testCase.pattern, testCase.testCases)
+	}
 }
 
 func Test_Utils_GetTrimmedParam(t *testing.T) {
 	t.Parallel()
-	res := GetTrimmedParam("*")
-	utils.AssertEqual(t, "*", res)
+	res := GetTrimmedParam("")
+	require.Equal(t, "", res)
+	res = GetTrimmedParam("*")
+	require.Equal(t, "*", res)
 	res = GetTrimmedParam(":param")
-	utils.AssertEqual(t, "param", res)
+	require.Equal(t, "param", res)
 	res = GetTrimmedParam(":param1?")
-	utils.AssertEqual(t, "param1", res)
+	require.Equal(t, "param1", res)
 	res = GetTrimmedParam("noParam")
-	utils.AssertEqual(t, "noParam", res)
+	require.Equal(t, "noParam", res)
+}
+
+func Test_Utils_RemoveEscapeChar(t *testing.T) {
+	t.Parallel()
+	res := RemoveEscapeChar(":test\\:bla")
+	require.Equal(t, ":test:bla", res)
+	res = RemoveEscapeChar("\\abc")
+	require.Equal(t, "abc", res)
+	res = RemoveEscapeChar("noEscapeChar")
+	require.Equal(t, "noEscapeChar", res)
+}
+
+func Benchmark_Utils_RemoveEscapeChar(b *testing.B) {
+	b.ReportAllocs()
+	b.ResetTimer()
+	var res string
+	for n := 0; n < b.N; n++ {
+		res = RemoveEscapeChar(":test\\:bla")
+	}
+
+	require.Equal(b, ":test:bla", res)
 }
 
 // go test -race -run Test_Path_matchParams
 func Benchmark_Path_matchParams(t *testing.B) {
-	type testparams struct {
-		url          string
-		params       []string
-		match        bool
-		partialCheck bool
-	}
 	var ctxParams [maxParams]string
-	benchCase := func(r string, cases []testparams) {
-		parser := parseRoute(r)
-		for _, c := range cases {
+	benchCaseFn := func(testCollection routeCaseCollection) {
+		parser := parseRoute(testCollection.pattern)
+		for _, c := range testCollection.testCases {
 			var matchRes bool
 			state := "match"
 			if !c.match {
 				state = "not match"
 			}
-			t.Run(r+" | "+state+" | "+c.url, func(b *testing.B) {
+			t.Run(testCollection.pattern+" | "+state+" | "+c.url, func(b *testing.B) {
 				for i := 0; i <= b.N; i++ {
 					if match := parser.getMatch(c.url, c.url, &ctxParams, c.partialCheck); match {
-						// Get params from the original path
+						// Get testCases from the original path
 						matchRes = true
 					}
 				}
-				utils.AssertEqual(t, c.match, matchRes, fmt.Sprintf("route: '%s', url: '%s'", r, c.url))
+				require.Equal(t, c.match, matchRes, fmt.Sprintf("route: '%s', url: '%s'", testCollection.pattern, c.url))
 				if matchRes && len(c.params) > 0 {
-					utils.AssertEqual(t, c.params[0:len(c.params)-1], ctxParams[0:len(c.params)-1], fmt.Sprintf("route: '%s', url: '%s'", r, c.url))
+					require.Equal(t, c.params[0:len(c.params)-1], ctxParams[0:len(c.params)-1], fmt.Sprintf("route: '%s', url: '%s'", testCollection.pattern, c.url))
 				}
 			})
-
 		}
 	}
-	benchCase("/api/:param/fixedEnd", []testparams{
-		{url: "/api/abc/fixedEnd", params: []string{"abc"}, match: true},
-		{url: "/api/abc/def/fixedEnd", params: nil, match: false},
-	})
-	benchCase("/api/v1/:param/*", []testparams{
-		{url: "/api/v1/entity", params: []string{"entity", ""}, match: true},
-		{url: "/api/v1/entity/", params: []string{"entity", ""}, match: true},
-		{url: "/api/v1/entity/1", params: []string{"entity", "1"}, match: true},
-		{url: "/api/v", params: nil, match: false},
-		{url: "/api/v2", params: nil, match: false},
-		{url: "/api/v1/", params: nil, match: false},
-	})
-	benchCase("/api/v1/:param", []testparams{
-		{url: "/api/v1/entity", params: []string{"entity"}, match: true},
-		{url: "/api/v1/entity/8728382", params: nil, match: false},
-		{url: "/api/v1", params: nil, match: false},
-		{url: "/api/v1/", params: nil, match: false},
-	})
-	benchCase("/api/v1", []testparams{
-		{url: "/api/v1", params: []string{}, match: true},
-		{url: "/api/v2", params: nil, match: false},
-	})
-	benchCase("/api/v1/:param/*", []testparams{
-		{url: "/api/v1/entity", params: []string{"entity", ""}, match: true},
-		{url: "/api/v1/entity/", params: []string{"entity", ""}, match: true},
-		{url: "/api/v1/entity/1", params: []string{"entity", "1"}, match: true},
-		{url: "/api/v", params: nil, match: false},
-		{url: "/api/v2", params: nil, match: false},
-		{url: "/api/v1/", params: nil, match: false},
-	})
+
+	for _, testCollection := range benchmarkCases {
+		benchCaseFn(testCollection)
+	}
+}
+
+// go test -race -run Test_RoutePatternMatch
+func Benchmark_RoutePatternMatch(t *testing.B) {
+	benchCaseFn := func(testCollection routeCaseCollection) {
+		for _, c := range testCollection.testCases {
+			// skip all cases for partial checks
+			if c.partialCheck {
+				continue
+			}
+			var matchRes bool
+			state := "match"
+			if !c.match {
+				state = "not match"
+			}
+			t.Run(testCollection.pattern+" | "+state+" | "+c.url, func(b *testing.B) {
+				for i := 0; i <= b.N; i++ {
+					if match := RoutePatternMatch(c.url, testCollection.pattern); match {
+						// Get testCases from the original path
+						matchRes = true
+					}
+				}
+				require.Equal(t, c.match, matchRes, fmt.Sprintf("route: '%s', url: '%s'", testCollection.pattern, c.url))
+			})
+		}
+	}
+
+	for _, testCollection := range benchmarkCases {
+		benchCaseFn(testCollection)
+	}
 }
