@@ -2,6 +2,7 @@ package fiber
 
 import (
 	"crypto/tls"
+	"errors"
 	"fmt"
 	"net"
 	"os"
@@ -9,39 +10,46 @@ import (
 	"runtime"
 	"strconv"
 	"strings"
+	"sync/atomic"
 	"time"
 
 	"github.com/valyala/fasthttp/reuseport"
+
+	"github.com/gofiber/fiber/v3/log"
 )
 
 const (
 	envPreforkChildKey = "FIBER_PREFORK_CHILD"
 	envPreforkChildVal = "1"
+	sleepDuration      = 100 * time.Millisecond
 )
 
 var (
 	testPreforkMaster = false
+	testOnPrefork     = false
 )
 
-// IsChild determines if the current process is a result of Prefork
+// IsChild determines if the current process is a child of Prefork
 func IsChild() bool {
 	return os.Getenv(envPreforkChildKey) == envPreforkChildVal
 }
 
 // prefork manages child processes to make use of the OS REUSEPORT or REUSEADDR feature
-func (app *App) prefork(addr string, tlsConfig *tls.Config) (err error) {
+func (app *App) prefork(addr string, tlsConfig *tls.Config, cfg ListenConfig) error {
+	var ln net.Listener
+	var err error
+
 	// 👶 child process 👶
 	if IsChild() {
 		// use 1 cpu core per child process
 		runtime.GOMAXPROCS(1)
-		var ln net.Listener
 		// Linux will use SO_REUSEPORT and Windows falls back to SO_REUSEADDR
 		// Only tcp4 or tcp6 is supported when preforking, both are not supported
-		if ln, err = reuseport.Listen("tcp4", addr); err != nil {
-			if !app.config.DisableStartupMessage {
-				time.Sleep(100 * time.Millisecond) // avoid colliding with startup message
+		if ln, err = reuseport.Listen(cfg.ListenerNetwork, addr); err != nil {
+			if !cfg.DisableStartupMessage {
+				time.Sleep(sleepDuration) // avoid colliding with startup message
 			}
-			return fmt.Errorf("prefork: %v", err)
+			return fmt.Errorf("prefork: %w", err)
 		}
 		// wrap a tls config around the listener if provided
 		if tlsConfig != nil {
@@ -50,6 +58,13 @@ func (app *App) prefork(addr string, tlsConfig *tls.Config) (err error) {
 
 		// kill current child proc when master exits
 		go watchMaster()
+
+		// prepare the server for the start
+		app.startupProcess()
+
+		if cfg.ListenerAddrFunc != nil {
+			cfg.ListenerAddrFunc(ln.Addr())
+		}
 
 		// listen for incoming connections
 		return app.server.Serve(ln)
@@ -61,14 +76,18 @@ func (app *App) prefork(addr string, tlsConfig *tls.Config) (err error) {
 		err error
 	}
 	// create variables
-	var max = runtime.GOMAXPROCS(0)
-	var childs = make(map[int]*exec.Cmd)
-	var channel = make(chan child, max)
+	max := runtime.GOMAXPROCS(0)
+	childs := make(map[int]*exec.Cmd)
+	channel := make(chan child, max)
 
 	// kill child procs when master exits
 	defer func() {
 		for _, proc := range childs {
-			_ = proc.Process.Kill()
+			if err := proc.Process.Kill(); err != nil {
+				if !errors.Is(err, os.ErrProcessDone) {
+					log.Errorf("prefork: failed to kill child: %v", err)
+				}
+			}
 		}
 	}()
 
@@ -77,8 +96,7 @@ func (app *App) prefork(addr string, tlsConfig *tls.Config) (err error) {
 
 	// launch child procs
 	for i := 0; i < max; i++ {
-		/* #nosec G204 */
-		cmd := exec.Command(os.Args[0], os.Args[1:]...)
+		cmd := exec.Command(os.Args[0], os.Args[1:]...) //nolint:gosec // It's fine to launch the same process again
 		if testPreforkMaster {
 			// When test prefork master,
 			// just start the child process with a dummy cmd,
@@ -92,8 +110,9 @@ func (app *App) prefork(addr string, tlsConfig *tls.Config) (err error) {
 		cmd.Env = append(os.Environ(),
 			fmt.Sprintf("%s=%s", envPreforkChildKey, envPreforkChildVal),
 		)
+
 		if err = cmd.Start(); err != nil {
-			return fmt.Errorf("failed to start a child prefork process, error: %v", err)
+			return fmt.Errorf("failed to start a child prefork process, error: %w", err)
 		}
 
 		// store child process
@@ -101,15 +120,33 @@ func (app *App) prefork(addr string, tlsConfig *tls.Config) (err error) {
 		childs[pid] = cmd
 		pids = append(pids, strconv.Itoa(pid))
 
+		// execute fork hook
+		if app.hooks != nil {
+			if testOnPrefork {
+				app.hooks.executeOnForkHooks(dummyPid)
+			} else {
+				app.hooks.executeOnForkHooks(pid)
+			}
+		}
+
 		// notify master if child crashes
 		go func() {
 			channel <- child{pid, cmd.Wait()}
 		}()
 	}
 
+	// Run onListen hooks
+	// Hooks have to be run here as different as non-prefork mode due to they should run as child or master
+	app.runOnListenHooks(app.prepareListenData(addr, tlsConfig != nil, cfg))
+
 	// Print startup message
-	if !app.config.DisableStartupMessage {
-		app.startupMessage(addr, tlsConfig != nil, ","+strings.Join(pids, ","))
+	if !cfg.DisableStartupMessage {
+		app.startupMessage(addr, tlsConfig != nil, ","+strings.Join(pids, ","), cfg)
+	}
+
+	// Print routes
+	if cfg.EnablePrintRoutes {
+		app.printRoutesMessage()
 	}
 
 	// return error if child crashes
@@ -123,25 +160,33 @@ func watchMaster() {
 		// and waits for it to exit
 		p, err := os.FindProcess(os.Getppid())
 		if err == nil {
-			_, _ = p.Wait()
+			_, _ = p.Wait() //nolint:errcheck // It is fine to ignore the error here
 		}
-		os.Exit(1)
+		os.Exit(1) //nolint:revive // Calling os.Exit is fine here in the prefork
 	}
 	// if it is equal to 1 (init process ID),
 	// it indicates that the master process has exited
-	for range time.NewTicker(time.Millisecond * 500).C {
+	const watchInterval = 500 * time.Millisecond
+	for range time.NewTicker(watchInterval).C {
 		if os.Getppid() == 1 {
-			os.Exit(1)
+			os.Exit(1) //nolint:revive // Calling os.Exit is fine here in the prefork
 		}
 	}
 }
 
-var dummyChildCmd = "go"
+var (
+	dummyPid      = 1
+	dummyChildCmd atomic.Value
+)
 
 // dummyCmd is for internal prefork testing
 func dummyCmd() *exec.Cmd {
-	if runtime.GOOS == "windows" {
-		return exec.Command("cmd", "/C", dummyChildCmd, "version")
+	command := "go"
+	if storeCommand := dummyChildCmd.Load(); storeCommand != nil && storeCommand != "" {
+		command = storeCommand.(string) //nolint:forcetypeassert,errcheck // We always store a string in here
 	}
-	return exec.Command(dummyChildCmd, "version")
+	if runtime.GOOS == "windows" {
+		return exec.Command("cmd", "/C", command, "version")
+	}
+	return exec.Command(command, "version")
 }
